### 2025-05-17 #8210 @Kevin-Mattheus-Moerman
@HzaCode thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:

- [ ] Please address the paper compilation error. I think perhaps you simply need to add the affiliation field and try again. You can try compilation again here by calling: `@editorialbot generate pdf`.
