### 2025-05-17 #8210 @HzaCode
Thanks so much for your guidance and for helping with the initial steps for my submission. I appreciate you taking the time to review it.

Regarding your points:

1.  **DOI Issues:** I will carefully review the reference check you've highlighted and address any potential problems by amending the DOI entries in my `.bib` file.
2. **Paper Compilation Error:** Thank you for the suggestion regarding the `affiliation` field. As an independent developer, I may not have a formal institutional affiliation. I will add "Independent Developer" to this field as required and then try compiling again by calling `@editorialbot generate pdf`.

I also wanted to mention that the current version submitted here has undergone significant updates and improvements since any versions that might have been available. Therefore, I kindly request that the review be based on the specific software version **tagged and referenced in the `paper.md` for this JOSS submission.**

I understand about the 'waitlisted' label and the current backlog. Thank you for letting me know and for your team's efforts in managing it. I appreciate your patience and am happy to wait.

I'll work on these points now and update you here shortly with the bot commands.

Thanks again for your help!
