+++ paper/paper.bib
@@ -1,52 +1,95 @@
 @article{PubChem,
-  author    = {Kim, Sunghwan and Chen, Jie and Cheng, Tiejun and Gindulyte, Asta and He, Jia and He, Siqian and Li, Qingrong and Shoemaker, Benjamin A and Thiessen, Paul A and Yu, Bo and Zaslavsky, Leonid and Zhang, Jian and Bolton, Evan E},
-  title     = {PubChem 2023 update},
-  journal   = {Nucleic Acids Research},
-  volume    = {51},
-  number    = {D1},
-  pages     = {D1373--D1380},
-  year      = {2023},
-  month     = {jan},
-  issn      = {1362-4962},
-  doi       = {10.1093/nar/gkac956},
-  url       = {https://doi.org/10.1093/nar/gkac956}
+  author  = {Kim, Sunghwan and Chen, Jie and Cheng, Tiejun and Gindulyte, Asta and He, Jia and He, Siqian and Li, Qingliang and Shoemaker, Benjamin A and Thiessen, Paul A and Yu, Bo and Zaslavsky, Leonid and Zhang, Jian and Bolton, Evan E},
+  title   = {PubChem 2023 Update},
+  journal = {Nucleic Acids Research},
+  volume  = {51},
+  number  = {D1},
+  pages   = {D1373--D1380},
+  year    = {2023},
+  doi     = {10.1093/nar/gkac956},
+  url     = {https://doi.org/10.1093/nar/gkac956}
+}
+
+@article{Kim2016NAR,
+  author  = {Kim, Sunghwan and Thiessen, Paul A. and Bolton, Evan E. and Chen, Jie and Fu, Gang and Gindulyte, Asta and Han, Lianyi and He, Jane and He, Siqian and Shoemaker, Benjamin A. and Wang, Jiyao and Yu, Bo and Zhang, Jian and Bryant, Stephen H.},
+  title   = {PubChem Substance and Compound Databases},
+  journal = {Nucleic Acids Research},
+  volume  = {44},
+  number  = {D1},
+  pages   = {D1202--D1213},
+  year    = {2016},
+  doi     = {10.1093/nar/gkv951},
+  url     = {https://doi.org/10.1093/nar/gkv951}
+}
+
+@article{Kim2018JCheminform,
+  author  = {H{\"a}hnke, Volker D. and Kim, Sunghwan and Bolton, Evan E.},
+  title   = {PubChem Chemical Structure Standardization},
+  journal = {Journal of Cheminformatics},
+  volume  = {10},
+  pages   = {36},
+  year    = {2018},
+  doi     = {10.1186/s13321-018-0293-8},
+  url     = {https://doi.org/10.1186/s13321-018-0293-8}
+}
+
+@article{Kim2018PUGREST,
+  author  = {Kim, Sunghwan and Thiessen, Paul A. and Cheng, Tiejun and Yu, Bo and Bolton, Evan E.},
+  title   = {An Update on PUG-REST: RESTful Interface for Programmatic Access to PubChem},
+  journal = {Nucleic Acids Research},
+  volume  = {46},
+  number  = {W1},
+  pages   = {W563--W570},
+  year    = {2018},
+  doi     = {10.1093/nar/gky294},
+  url     = {https://doi.org/10.1093/nar/gky294}
 }
 
 @misc{RequestsCache,
-  author       = {Roman, Jordan W and others},
-  title        = {requests-cache: Transparent persistent cache for python requests},
-  howpublished = {\url{https://github.com/requests-cache/requests-cache}},
-  year         = {2024},
-  note         = {Accessed: 2025-04-15}
+  author = {Roman, Jordan W. and others},
[…]
