+++ paper/paper.md
@@ -10,47 +10,54 @@ tags:
   - cheminformatics
   - research software
   - data validation
-authors:
-  - name: Zhiang HE
+authors: 
+  - name: Zhiang He
     orcid: 0009-0009-0171-4578
-date: 15 April 2025 
-bibliography: paper.bib 
-repository: https://github.com/HzaCode/ChemInformant 
+affiliation: 
+date: 15 April 2025
+bibliography: paper.bib # Assuming your .bib file is named paper.bib
+repository: https://github.com/HzaCode/ChemInformant
 version: 1.1.5 
-license: MIT 
-doi: null 
+license: MIT
+doi:
 ---
 
-# Summary
+## Summary
 
-ChemInformant provides a **lightweight, highly user-friendly Pythonic interface** for accessing chemical compound data from the PubChem database [@PubChem]. Designed with simplicity and robustness in mind (`import ChemInformant as ci`), it shields researchers from the complexities of the PubChem API. Core advantages include **built-in automatic caching** via `requests-cache` [@RequestsCache] significantly enhancing performance and reproducibility, **structured data validation** using Pydantic [@Pydantic] models (`CompoundData`) ensuring data quality, **explicit handling of chemical name ambiguity** improving accuracy, and **optimized batch retrieval**. Critically, ChemInformant is built upon a minimal set of **modern, actively maintained core dependencies** (`requests` [@Requests], `requests-cache`, `Pydantic`), ensuring long-term stability and compatibility, addressing limitations found in direct API usage and offering improvements over aspects of existing tools that may rely on less modern or potentially less maintained foundations.
+ChemInformant provides a **lightweight, highly user-friendly Pythonic interface** for accessing chemical compound data from the PubChem database [@PubChem; @Kim2016NAR]. Designed with simplicity and robustness in mind (`import ChemInformant as ci`), it shields researchers from the complexities of the PubChem API. Core advantages include **built-in automatic caching** via `requests-cache` [@RequestsCache] significantly enhancing performance and reproducibility, **structured data validation** using Pydantic [@Pydantic] models (`CompoundData`) ensuring data quality, **explicit handling of chemical name ambiguity** improving accuracy, and **optimized batch retrieval**. Critically, ChemInformant is built upon a minimal set of **modern, actively maintained core dependencies** (`requests` [@Requests], `requests-cache`, `Pydantic`), ensuring long-term stability and compatibility. This addresses limitations found in direct API usage and offers improvements over aspects of existing tools that may rely on less modern or potentially less maintained foundations. ChemInformant aims to be a reliable and efficient tool for routine chemical information retrieval.
 
-# Statement of Need
+## Statement of Need
 
-Accessing PubChem [@PubChem] programmatically is essential for many chemistry-related research tasks, yet direct use of the PUG REST API involves significant overhead in request handling, response parsing, error management, and data validation. This complexity can slow down research and introduce errors.
+Accessing PubChem [@PubChem] programmatically is essential for many chemistry-related research tasks, yet direct use of the PUG REST API [@Kim2018JCheminform; @Kim2018PUGREST] involves significant overhead in request handling, response parsing, error management, and data validation. This complexity can slow down research and introduce errors.
 
 While existing libraries like PubChemPy [@PubChemPy] offer an interface to PubChem, `ChemInformant` was developed to specifically address several key limitations and provide a more modern, robust, and efficient experience tailored to common research workflows:
 
[…]
