+++ paper/paper.md
@@ -98,7 +98,7 @@ for ident, result in results.items():
 | :-------------------------- | :------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | :------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
 | **Persistent HTTP Cache**   | Built-in `requests-cache`, supporting SQLite/memory/Redis backends. Subsequent requests for cached data can achieve near-zero network traffic in repetitive analysis scenarios, significantly improving efficiency.| No global caching mechanism; each batch request re-accesses the network, resulting in an inability to leverage caching for acceleration during repeated retrievals, and fixed network overhead. |
 | **Runtime Data Validation** | Employs Pydantic `CompoundData` model for immediate validation of field types and missing data, actively raising errors.                                                               | Returns native Python objects (e.g., dictionaries) with no built-in mandatory validation mechanism; users are responsible for their own data checks.                                         |
-| **Chemical Name Disambiguation** | Automatically raises an `AmbiguousIdentifierError` exception; the exception object (`e`) includes an `e.cids` attribute listing all candidate CIDs, allowing developers to select one to proceed (e.g., `ci.info(e.cids[0])`). Enforces handling of ambiguity to prevent misuse of compounds. | Directly returns a list of results without erroring on ambiguity; developers must manually check the list length (e.g., `len(hits)`) and review the results, otherwise, it defaults to using the first item in the list, which can easily lead to incorrect selections. |
+| **Chemical Name Disambiguation** | Automatically raises an `AmbiguousIdentifierError` exception; the exception object (`e`) includes an `e.cids` attribute listing all candidate CIDs, allowing developers to select one to proceed (e.g., `ci.info(e.cids[0])`). Enforces handling of ambiguity to prevent misuse of compounds. |Directly returns a list of results without erroring on ambiguity, which can easily lead to incorrect selections. |
 | **Batch Retrieval**         | Single call to `ci.get_multiple_compounds(ids)`: supports mixed ID types, automatically handles classification, pagination (≤ 200 entries/batch), and rate limiting (≤ 5 reqs/sec); error entries are encapsulated as exception objects, while valid entries return complete `CompoundData`. | Requires users to manually split IDs by namespace, then loop through `get_properties` / `get_compounds`, and self-manage request rates (e.g., via `sleep`); if batch size exceeds limits or URL construction is incorrect, the entire batch request fails. |
 | **2-D Structure Drawing**   | Provides `ci.fig(cid)` method for single-line rendering via Pillow and Matplotlib; supports inline display in Jupyter Notebooks.                                                       | No built-in drawing functionality; only returns image URLs (PNG/SVG format), requiring users to manually download them for display.                                                              |
 | **Core Dependencies & Maintenance** | Core dependencies include `requests` 2.x, `requests-cache` 1.x, `pydantic` 2.x, all actively maintained versions; the CI package itself is also under continuous development and updates . | Core dependency is primarily `requests`; PCP's latest stable version was released in 2017, with low subsequent maintenance activity.                                                              |
