+++ paper/paper.md
@@ -15,46 +15,47 @@ authors:
     orcid: 0009-0009-0171-4578
 affiliation: 
 date: 15 April 2025
-bibliography: paper.bib # Assuming your .bib file is named paper.bib
+bibliography: paper.bib 
 repository: https://github.com/HzaCode/ChemInformant
 version: 1.1.5 
 license: MIT
 doi:
 ---
 
+
 ## Summary
 
-ChemInformant provides a **lightweight, highly user-friendly Pythonic interface** for accessing chemical compound data from the PubChem database [@PubChem; @Kim2016NAR]. Designed with simplicity and robustness in mind (`import ChemInformant as ci`), it shields researchers from the complexities of the PubChem API. Core advantages include **built-in automatic caching** via `requests-cache` [@RequestsCache] significantly enhancing performance and reproducibility, **structured data validation** using Pydantic [@Pydantic] models (`CompoundData`) ensuring data quality, **explicit handling of chemical name ambiguity** improving accuracy, and **optimized batch retrieval**. Critically, ChemInformant is built upon a minimal set of **modern, actively maintained core dependencies** (`requests` [@Requests], `requests-cache`, `Pydantic`), ensuring long-term stability and compatibility. This addresses limitations found in direct API usage and offers improvements over aspects of existing tools that may rely on less modern or potentially less maintained foundations. ChemInformant aims to be a reliable and efficient tool for routine chemical information retrieval.
+ChemInformant provides a **lightweight Pythonic interface** characterized by its ease of use for accessing chemical compound data from the PubChem database [@PubChem; @Kim2016NAR]. Designed to abstract the complexities of the PubChem API (`import ChemInformant as ci`), it facilitates streamlined data retrieval for researchers. Key distinguishing features include **built-in automatic caching** via `requests-cache` [@RequestsCache], which enhances performance and contributes to reproducibility; **structured data validation** using Pydantic [@Pydantic] models (`CompoundData`), promoting data quality; **explicit handling of chemical name ambiguity**, enhancing accuracy; and **optimized batch retrieval**. Furthermore, ChemInformant is built upon a minimal set of **modern, actively maintained core dependencies** (`requests` [@Requests], `requests-cache`, `Pydantic`), which contributes to its long-term stability and compatibility. This design addresses certain limitations associated with direct API utilization and offers enhancements over existing tools that may be built upon less contemporary or less actively maintained software foundations. ChemInformant is designed as a reliable and efficient tool for routine chemical information retrieval.
 
 ## Statement of Need
 
-Accessing PubChem [@PubChem] programmatically is essential for many chemistry-related research tasks, yet direct use of the PUG REST API [@Kim2018JCheminform; @Kim2018PUGREST] involves significant overhead in request handling, response parsing, error management, and data validation. This complexity can slow down research and introduce errors.
+Programmatic access to PubChem [@PubChem] is integral for numerous chemistry-related research tasks; however, direct utilization of the PUG REST API [@Kim2018JCheminform; @Kim2018PUGREST] can entail considerable overhead in request handling, response parsing, error management, and data validation. This inherent complexity can impede research progress and potentially introduce errors.
 
-While existing libraries like PubChemPy [@PubChemPy] offer an interface to PubChem, `ChemInformant` was developed to specifically address several key limitations and provide a more modern, robust, and efficient experience tailored to common research workflows:
+While existing libraries, such as PubChemPy [@PubChemPy], provide an interface to PubChem, `ChemInformant` was developed to address specific limitations and to offer a contemporary, robust, and efficient alternative tailored to common research workflows by focusing on the following aspects:
 
[…]
