+++ paper/paper.md
@@ -104,7 +104,7 @@ The features detailed in Table 1, particularly when contrasted with the baseline
 | **Automatic Persistent HTTP Caching**<br/>(Efficiency in Repetitive Data Access) | Implements built-in, transparent HTTP caching via `requests-cache`, configurable (e.g., SQLite backend, 7-day expiry, caches 404 responses). This significantly accelerates repetitive chemical data lookups by minimizing network requests. | Lacks a global persistent HTTP response caching mechanism; primarily relies on instance-level attribute caching for chemical properties. This approach is less effective for performance optimization across sessions or during large dataset processing. |
 | **Runtime Data Validation**<br/>(Integrity of Chemical Information)          | Employs Pydantic models (`CompoundData`) for **strict structural and data type validation** of retrieved chemical data (e.g., SMILES, MW, InChI). This ensures data integrity for subsequent modeling or analysis tasks.      | Does not incorporate an explicit client-side data validation layer for chemical attributes. Consequently, there is a potential for inconsistent data types or structures to propagate into downstream processes.                 |
 | **Chemical Name Disambiguation**<br/>(Accuracy of Identifier Mapping)        | **Proactively raises an `AmbiguousIdentifierError`** with candidate CIDs for ambiguous chemical names. This mechanism guides users towards accurate entity selection, crucial for maintaining dataset quality.           | Typically returns a list of matching chemical entities, necessitating user-side logic for disambiguation. This may introduce a higher propensity for errors in automated workflows if not explicitly handled.               |
-| **Optimized Batch Retrieval**<br/>(Efficiency for Multiple Entities/Properties) | Provides `get_multiple_compounds` for **efficient retrieval of comprehensive chemical profiles** for multiple compounds, thereby minimizing API calls. This is advantageous for screening large chemical datasets.        | The `get_properties` function allows batch retrieval of a specified, limited set of core chemical properties. Fetching diverse property types for numerous compounds may require iterative API calls or custom aggregation logic. |
+| **Optimized Batch Retrieval**<br/>(Efficiency for Multiple Entities/Properties) | Single call to `ci.get_multiple_compounds(ids)`: supports mixed ID types, automatically handles classification, pagination (≤ 200 entries/batch), and rate limiting (≤ 5 reqs/sec); error entries are encapsulated as exception objects, while valid entries return complete `CompoundData`. | Requires users to manually split IDs by namespace, then loop through `get_properties` / `get_compounds`, and self-manage request rates (e.g., via `sleep`); if batch size exceeds limits or URL construction is incorrect, the entire batch request fails. |
 | **Integrated 2D Chemical Structure Plotting**<br/>(Rapid Visual Inspection)    | Includes a convenience function, `ci.fig()` (requires `matplotlib`, `Pillow`), for direct 2D visualization of chemical structures, facilitating quick data exploration and reporting.                              | Lacks a built-in, direct display function for chemical structures; provides image URLs that require user-side processing for visualization.                                                                                |
 | **Core Dependency Status**<br/>(Long-term Sustainability and Ecosystem Integration) | Relies on **modern, actively maintained libraries** (`requests`, `requests-cache`, `Pydantic`), ensuring long-term viability and robust integration with other contemporary cheminformatics tools.                | The library (v1.0.4) has not received significant updates since its 2017 release. This raises potential concerns regarding long-term maintenance and compatibility within evolving cheminformatics pipelines.               |
 
@@ -112,3 +112,5 @@ The features detailed in Table 1, particularly when contrasted with the baseline
 ## Acknowledgement
 
 We acknowledge the PubChem database [@PubChem] as the primary data source. This work utilizes a series of open-source Python libraries for tasks including data requests, caching, validation, as well as optional image processing and data visualization.
+
+
