### 2025-05-22 #8210 @HzaCode
Hi @espottesmith, thank you for handling the review.

Here are several potential reviewers I would like to suggest, based on their expertise in cheminformatics, scientific software development, and experience with technologies relevant to this submission (e.g., API client design, caching, validation frameworks):

* **@cthoyt** – Active in biomedical knowledge graph tooling and open-source data infrastructure, with extensive work in ontology and cheminformatics-related tools.
* **[[@Sulstice](https://github.com/Sulstice)](https://github.com/Sulstice)** – Maintains repositories involving molecule-based computations and scientific Python packaging.
* **[[@andrewtarzia](https://github.com/andrewtarzia)](https://github.com/andrewtarzia)** – Works on molecular structure modeling and cheminformatics toolkits using Python.
* **[[@RMeli](https://github.com/RMeli)](https://github.com/RMeli)** – Strong background in physical chemistry and cheminformatics, contributes to multiple scientific software libraries.

These names were identified through the JOSS reviewer search tool using the keywords “cheminformatics” and programming language “Python.”
