### 2025-05-22 #8210 @HzaCode
Regarding 10.1201/9781584889304-33, we found that this DOI refers to a specific book chapter published by CRC Press, and it is not an official publication of the Python Software Foundation. To ensure the citation accurately reflects the source being referenced, we have chosen not to include this DOI for the Python Language Reference.
