+++ README.md
@@ -25,7 +25,6 @@ ChemInformant <img src="https://github.com/HzaCode/ChemInformant/blob/main/image
 [![PyPI version](https://img.shields.io/pypi/v/ChemInformant.svg)](https://pypi.org/project/ChemInformant/) [![PyPI - Python Version](https://img.shields.io/pypi/pyversions/ChemInformant.svg)](https://pypi.org/project/ChemInformant/)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![Build Status](https://img.shields.io/github/actions/workflow/status/HzaCode/ChemInformant/tests.yml?label=Build)](https://github.com/HzaCode/ChemInformant/actions/workflows/tests.yml) [![Code Quality: Flake8](https://img.shields.io/github/actions/workflow/status/HzaCode/ChemInformant/lint.yml?label=Code%20Quality%3A%20Flake8)](https://github.com/HzaCode/ChemInformant/actions/workflows/lint.yml) [![PyPI - License](https://img.shields.io/pypi/l/ChemInformant.svg)](https://github.com/HzaCode/ChemInformant/blob/main/LICENSE) [![Docs](https://img.shields.io/badge/Docs-Read_Online-blue?style=flat-square&logo=book&logoColor=white)](https://hzacode.github.io/ChemInformant)[![codecov](https://codecov.io/gh/HzaCode/ChemInformant/branch/main/graph/badge.svg)](https://codecov.io/gh/HzaCode/ChemInformant) 
-[![arXiv](https://img.shields.io/badge/arXiv-preprint-B31B1B.svg)](< in preparation >)
 
 
 
