### 2025-05-22 #8210 @cthoyt
does this pass JOSS's [substantial scholarly effort](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort) criterion? It appears the software repository is only about two months old
