### 2025-05-23 #8210 @espottesmith
@cthoyt I judge this code to represent a substantial scholarly effort. As you note, there has been a flurry of recent effort, but the work was started earlier. The code is also substantial (several thousand lines of Python code).

Does this satisfy your query? And, if so, would you be willing to serve as a reviewer?
