+++ README.md
@@ -2,6 +2,15 @@ ChemInformant <img src="https://github.com/HzaCode/ChemInformant/blob/main/image
 ---  
 
 
+<div align="left">
+  <a href="https://pepy.tech/project/cheminformant" target="_blank" title="View download stats">
+    <img
+      src="https://static.pepy.tech/personalized-badge/cheminformant?period=total&units=abbreviation&left_color=grey&right_color=blue&left_text=Downloads"
+      alt="Total Downloads"
+      height="35">
+  </a>
+</div>
+
 
 
 
