### 2025-06-04 #8341 @editorialbot
[REVIEW]: ChemInformant: A Modern, Lightweight Python Client for PubChem with Robust Caching and Validation
**Submitting author:** @HzaCode (<a href="http://orcid.org/0009-0009-0171-4578">Zhiang HEZHIANG</a>)
**Repository:** https://github.com/HzaCode/ChemInformant
**Branch with paper.md** (empty if default branch): main
**Version:** v2.4.0
