+++ examples/basic_usage.py
@@ -1,11 +1,9 @@
-# examples/basic_usage_en.py
-
 import sys
 import os
 import time
 
 try:
-    # --- Specify the alias using the 'as ci' keyword ---
+    # Specify an alias for convenience
     import ChemInformant as ci
 except ImportError as e:
     print(f"Import Error: {e}")
@@ -15,15 +13,15 @@ except ImportError as e:
     )
     sys.exit(1)
 
-# --- Optional: Configure Cache ---
-# Default cache is 'pubchem_cache.sqlite' in the current directory, expires after 7 days.
-# If you want to change settings (e.g., use memory cache, expire after 1 hour), uncomment the line below:
+# Optional: Configure caching
+# Default cache file is 'pubchem_cache.sqlite' in the current directory, expires after 7 days.
+# To change settings (e.g., use memory cache, expire after 1 hour), uncomment the line below:
 # ci.setup_cache(backend='memory', expire_after=3600)
 # print(f"Using cache file (default): {os.path.abspath('pubchem_cache.sqlite')}") # Show default path
 
 
 def print_separator(title=""):
-    """Prints a formatted separator line."""
+    """Prints a formatted separator line to improve output readability."""
     width = 80
     if title:
         print(f"\n--- {title} {'-' * (width - len(title) - 5)}")
@@ -32,64 +30,73 @@ def print_separator(title=""):
 
 
 def demonstrate_single_compound(compound_identifier):
-    """Gets and prints details for a single compound, using the 'ci' alias."""
+    """Gets and prints detailed information for a single compound."""
     print_separator(
         f"Single Query: {compound_identifier} (Type: {type(compound_identifier).__name__})"
     )
 
     try:
-        # === Call functions using the ci alias ===
+        # Call functions using the 'ci' alias
         print("Attempting to get individual properties:")
+        # Note: ci.cid() aims to find a single, best-matching CID.
+        # For more complex cases, like ambiguity, ci.info() might provide more detailed information.
+        print(f"  (Note: ci.cid() attempts to find a single, best-matching CID. More complex cases like ambiguity might be detailed by ci.info().)")
         start_time_cid = time.perf_counter()
-        compound_cid = ci.cid(compound_identifier)  # Calling ci.cid()
+        compound_cid = ci.cid(compound_identifier)
         end_time_cid = time.perf_counter()
         print(f"  - Calling ci.cid() took: {end_time_cid - start_time_cid:.4f}s")
 
         if compound_cid is None:
             print(
-                f"  - Could not definitively resolve CID for '{compound_identifier}' (maybe not found or ambiguous)"
+                f"  - ci.cid() did not return a definitive CID for '{compound_identifier}'."
             )
-            # Try calling info() to get a more specific error
+            print("  - Attempting diagnostic ci.info() call for more details...")
             try:
+                # The call to ci.info() here is primarily to leverage specific errors it might throw
+                # (like NotFoundError, AmbiguousIdentifierError) to help understand why the ci.cid() call failed.
+                # The result of this specific ci.info() call is not directly used for subsequent processing here.
                 ci.info(compound_identifier)
-            except (ci.NotFoundError, ci.AmbiguousIdentifierError) as specific_error:
-                print(f"  - Specific error: {specific_error}")
+                print("  - Diagnostic ci.info() call did not raise a specific error, but CID remains unresolved.")
+            except (ci.NotFoundError, ci.AmbiguousIdentifierError) as specific_error: # Keep handling for AmbiguousIdentifierError
+                print(f"  - Specific error from diagnostic ci.info() call: {specific_error}")
             except Exception as unknown_error:
[…]
