### 2025-06-05 2fd4e30
Delete minimal_batch_test.py
