+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ChemInformant"
-version = "1.2.0"
+version = "1.2.1"
 authors = [
  
   { name="Ang", email="ang@hezhiang.com" },
