+++ CITATION.cff
@@ -1,12 +1,12 @@
 cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 authors:
-  - family-names: "HE"
+  - family-names: "He"
     given-names: "Zhiang"
     orcid: "https://orcid.org/0009-0009-0171-4578"
-title: "ChemInformant: A Modern, Lightweight Python Client for PubChem with Robust Caching and Validation"
-version: "1.1.5"
-date-released: 2025
+title: "ChemInformant"
+version: 1.2.1
+date-released: 2024-05-20
 url: "https://github.com/HzaCode/ChemInformant"
 license: MIT
 keywords:
