### 2025-06-06 2eded5a
Update CITATION.cff
