+++ CITATION.cff
@@ -6,7 +6,7 @@ authors:
     orcid: "https://orcid.org/0009-0009-0171-4578"
 title: "ChemInformant"
 version: 1.2.1
-date-released: 2024-05-20
+date-released: 2025
 url: "https://github.com/HzaCode/ChemInformant"
 license: MIT
 keywords:
