### 2025-06-24 HzaCode/ChemInformant#1 @cthoyt
[REVIEW] Basic usage in README is too cluttered
RE: https://github.com/openjournals/joss-reviews/issues/8341

The basic usage section (https://github.com/HzaCode/ChemInformant?tab=readme-ov-file#basic-usage) in the readme contains tons of visual noise

- https://github.com/HzaCode/ChemInformant/blob/main/README.md?plain=1#L91 don't show any optional code at all in basic usage. e.g., the caching should be in documentation (or should just automatically happen, if you really think it's important)
- https://github.com/HzaCode/ChemInformant/blob/main/README.md?plain=1#L98 why not use the fact that you're in markdown to your advantage? use markdown that has narrative interspersed with code, rather than shoving as many possible comments into the code itself
- https://github.com/HzaCode/ChemInformant/blob/main/README.md?plain=1#L100 why do you need to nest all code in try/except? The README should contain the most important functionality to get going, error handling seems like something that should be part of the documentation and not on the front page
- https://github.com/HzaCode/ChemInformant/blob/main/README.md?plain=1#L191 consider using [pystow](https://github.com/cthoyt/pystow) for automating choosing a consistent path for caching
- https://github.com/HzaCode/ChemInformant/blob/main/README.md?plain=1#L202 use a jupyter notebook to show off displaying the compound image, I don't think any user would do this from a REPL or inside non-interactive code
- https://github.com/HzaCode/ChemInformant/blob/main/README.md?plain=1#L216 this kind of documentation should be in a reference, inside the docs, but not on the README. I'd suggest watching this excellent talk on how to structure documentation: https://www.youtube.com/watch?v=t4vKPhjcMZg
