### 2025-06-24 HzaCode/ChemInformant#2 @cthoyt
[REVIEW] Use extras in pyproject.toml for extra features
RE: https://github.com/openjournals/joss-reviews/issues/8341

In the README's installation documentation, it asks users to manually install some dependencies needed for generating images

https://github.com/HzaCode/ChemInformant/blob/2eded5a0599152a39a191277336411a2ff7a38bc/README.md?plain=1#L76

There's a better way to do this, where you explicitly encode this as a group of "optional" dependencies - see https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-optional-dependencies on how to add an `[project.optional-dependencies]` which would look something like this in your `pyproject.toml`:

```toml
[project.optional-dependencies]
draw = [
    "matplotlib",
    "pillow",
]
```

then you do `pip install .[draw]`
