### 2025-06-24 HzaCode/ChemInformant#4 @cthoyt
[REVIEW] Provide benchmark
RE: https://github.com/openjournals/joss-reviews/issues/8341

Since the documentation and JOSS submission state the key feature of this package is its caching mechanism, can you provide a benchmark that shows in which cases this is useful / more performant that making my own calls to the PubChem API (e.g., w/ requests by itself)?

Ideally such a benchmark would be fully automated (i.e., anyone can run it from their own machine) and produces a nice comparison chart between pure requests calls, and calls made with ChemInformant
