### 2025-06-24 HzaCode/ChemInformant#5 @cthoyt
[REVIEW] Drop pydantic v1 support
RE: https://github.com/openjournals/joss-reviews/issues/8341

Pydantic V1 is long past end of life, so keeping support for it is confusing and causes annoying warnings for downstream users
