### 2025-06-24 HzaCode/ChemInformant#6 @cthoyt
[REVIEW] Remove patronizing language about ease of use
RE: https://github.com/openjournals/joss-reviews/issues/8341

This might be a matter of taste, but I do not like when software packages make claims about ease of use and simplicity. I feel patronized as a reader/user by this kind of language. Also, in general, such claims are very easy to prove false. For example, in the following section:

https://github.com/HzaCode/ChemInformant/blob/2eded5a0599152a39a191277336411a2ff7a38bc/paper/paper.md?plain=1#L41

I would argue that the interface is (in its current state) neither Pythonic nor intuitive (more detail in https://github.com/HzaCode/ChemInformant/issues/3).

I would suggest removing all such language both from the software documentation and associated manuscript (even if you address the underlying issues).
