### 2025-06-24 HzaCode/ChemInformant#7 @cthoyt
[REVIEW] Add better description about state of the field
RE: https://github.com/openjournals/joss-reviews/issues/8341

A cursory search on PyPI shows that there are many packages for interacting with PubChem: https://pypi.org/search/?q=pubchem

It would be valuable to expand the section about the current state of the field. What are the important features that other packages have? Does ChemInformant also have these features, or is it missing some things? This will also help better streamline the "statement of need section" since you can (then) say "here are the features that don't exist elsewhere, and this is why I needed to write my own package that has them"
