### 2025-06-24 HzaCode/ChemInformant#8 @cthoyt
[REVIEW] Improve statement of need
RE: https://github.com/openjournals/joss-reviews/issues/8341

On top of https://github.com/HzaCode/ChemInformant/issues/7, it's not yet clear what's the statement of need.

https://github.com/HzaCode/ChemInformant/blob/2eded5a0599152a39a191277336411a2ff7a38bc/paper/paper.md?plain=1#L37

I don't think this is an issue for other packages, and isn't itself a need to create a new one

https://github.com/HzaCode/ChemInformant/blob/2eded5a0599152a39a191277336411a2ff7a38bc/paper/paper.md?plain=1#L38

this is probably the most important thing you should focus on. But you have to go further than just saying that you provide caching. Why does this require making a whole new library? Why not just apply existing caching functionality on top of something that exists, e.g., using https://github.com/python-cachier/cachier?

https://github.com/HzaCode/ChemInformant/blob/2eded5a0599152a39a191277336411a2ff7a38bc/paper/paper.md?plain=1#L39

this is nice, but you can make a much better point here. The benefit of pydantic is that it's a generic and well-known validation system that is effectively declarative in python code

https://github.com/HzaCode/ChemInformant/blob/2eded5a0599152a39a191277336411a2ff7a38bc/paper/paper.md?plain=1#L40

this is a weak point on its own without clarification. Why is error handling notable when working with PubChem? Have other packages not addressed this?

https://github.com/HzaCode/ChemInformant/blob/2eded5a0599152a39a191277336411a2ff7a38bc/paper/paper.md?plain=1#L41

please delete this point completely, see #6
