### 2025-06-24 #8341 @cthoyt
@espottesmith @HzaCode I took a first pass at reviewing the code and paper. I'd love to see some improvements to the code to make the library more usable (specifics in issues on the repo), updates to the paper to make a better case for the existence of this library vs. the many other existing (Python) PubChem API wrappers (also specific suggestions in issues on the repo), and some general writing improvements. Cheers
