### 2025-06-27 #8341 @LIVazquezS
Dear Authors,

Here are some of my thoughts.

I appreciate the initiative to create software that can automatically wrap data from the PubChem database. Particularly in view that PubChemPy is not continuously maintained. However, I believe that, in its current state, Cheminformant does not provide all capabilities that PubChemPy or other alternatives in R (See PubChemR ([10.32614/RJ-2024-031](https://doi.org/10.32614/RJ-2024-031)).

I understand the authors' desire to create a tool that can be easily integrated into automatic pipelines. However, I failed to see how this can be done by providing specific options to retrieve particular properties or allowing users to save the requested data automatically.

## General Checks

### Functionalities

As mentioned above, the paper claims to be an alternative to PubChemPy. However, this evaluation is done in specific aspects while ignoring others, such as the obtention of multiple properties of a compound. In the current version of the code, it is only possible to retrieve basic properties (i.e. molecular weight, formula, IUPAC name). The code becomes more applicable if other properties can also be retrieved.

Similarly, increasing the capabilities to save batches of compounds, either by using Pandas or SQL files, would enhance the applicability of the code.

### Example Usage

The current example shows the basics of the functionalities of the code. However, a more advanced example that shows the integrability of cheminformant in pipelines or its use in combination with existing cheminformatics tools is, in my opinion, missing.

In addition, the example could be better integrated into the documentation.

### Functional Documentation

The documentation requires further refinement to improve its usability. Clear sections explaining installation, contributing, and primary functions are needed. I recommend that the authors move some of the information from the README file on GitHub to the documentation.

## Software paper

### Statement of need

The authors need to improve the statement of need. In its current form, the authors have focused too much on providing a tool that can easily retrieve information from PubChem using modern tools. However, there is no justification for what data can be retrieved, how it can be handled, and so on.

### State of the field

The authors need to conduct a more thorough review of the current state of the art. There is no mention of tools in other languages, such as R or Java, which provide multiple other capabilities. In addition, the authors do not mention other databases, such as Enamine or ChemSpider, which could be integrated into the package later.

### Quality of writing

The paper has a few typos:

- Page 2: The figure is not correctly referenced.
- Figure 1: The name of the package is incomplete
- Table 1: It has formatting problems.

Also, additional details about the presented example should be provided. It is unclear what the output of the Python code displayed is.

## Overall

I applaud the initiative of the authors in creating a tool that uses state-of-the-art packages for retrieving information from PubChem. This is not an easy task, and it is a need for the field. However, I believe that the code requires considerable improvements to be accepted in JOSS. Nevertheless, I saw potential in this code to become a standard tool for the cheminformatics community.
