### 2025-06-27 #8341 @HzaCode
@LIVazquezS,
Thanks so much for this detailed feedback. It's incredibly helpful.
Your points, along with the earlier suggestions from @cthoyt, give me a clear path forward for improving the project. I'm getting started on these changes now.
Thanks again for your time and thoughtful comments.
