+++ tests/test_api.py
@@ -1,1102 +1,0 @@
-import pytest
-import requests
-# import requests_cache # Removed unused
-# from pydantic import HttpUrl # Not directly used in tests, but CompoundData might use it
-from unittest.mock import MagicMock, patch, ANY  # Re-add ANY
-# import io # Removed unused
-import sys # Ensure sys is imported
-import re  # Add import for re module
-
-from ChemInformant import (
-    info,
-    cid,
-    cas,
-    unii,
-    form,
-    wgt,
-    smi,
-    iup,
-    dsc,
-    syn,
-    get_multiple_compounds,
-    setup_cache,
-    CompoundData,
-    NotFoundError,
-    AmbiguousIdentifierError,
-    fig,
-)
-from requests.exceptions import ConnectionError
-
-MOCK_CID_ASPIRIN = 2244
-MOCK_CID_ETHANOL = 702
-MOCK_CID_WATER = 962
-MOCK_CID_AMBIGUOUS_1 = 1001
-MOCK_CID_AMBIGUOUS_2 = 1002
-MOCK_CID_NO_CAS = 1003
-MOCK_CID_NO_WEIGHT = 1004
-MOCK_CID_FETCH_ERROR_CAS_UNII = 1005
-MOCK_CID_FETCH_ERROR_PROPS = 1006
-MOCK_CID_FETCH_ERROR_DESC = 1007
-MOCK_CID_FETCH_ERROR_SYN = 1008
-MOCK_CID_INVALID_DATA = 1009
-
-MOCK_NAME_ASPIRIN = "Aspirin"
-MOCK_NAME_ETHANOL = "Ethanol"
-MOCK_NAME_WATER = "Water"
-MOCK_NAME_AMBIGUOUS = "AmbiguousDrug"
-MOCK_NAME_NOTFOUND = "NotFoundCompound"
-MOCK_NAME_NOCAS = "NoCasCompound"
-MOCK_NAME_NOWEIGHT = "NoWeightCompound"
-MOCK_NAME_FETCH_ERROR_CAS_UNII = "FetchErrorCasUniiCompound"
-MOCK_NAME_FETCH_ERROR_PROPS = "FetchErrorPropsCompound"
-MOCK_NAME_FETCH_ERROR_DESC = "FetchErrorDescCompound"
-MOCK_NAME_FETCH_ERROR_SYN = "FetchErrorSynCompound"
-MOCK_NAME_INVALID_DATA = "InvalidDataCompound"
-EXPECTED_ASPIRIN_URL = f"https://pubchem.ncbi.nlm.nih.gov/compound/{MOCK_CID_ASPIRIN}"
-
-
-@pytest.fixture
-def mock_api_helpers(mocker):
-    mock_get_cids = mocker.patch(
-        "ChemInformant.cheminfo_api.api_helpers.get_cids_by_name"
-    )
-    mock_get_cas_unii = mocker.patch(
-        "ChemInformant.cheminfo_api.api_helpers.get_cas_unii"
-    )
-    mock_get_props = mocker.patch(
-        "ChemInformant.cheminfo_api.api_helpers.get_additional_properties"
-    )
-    mock_get_desc = mocker.patch(
-        "ChemInformant.cheminfo_api.api_helpers.get_compound_description"
-    )
-    mock_get_syns = mocker.patch(
-        "ChemInformant.cheminfo_api.api_helpers.get_all_synonyms"
-    )
-
-    mock_get_batch_props = mocker.patch(
-        "ChemInformant.cheminfo_api.api_helpers.get_batch_properties"
-    )
-    mock_get_batch_syns = mocker.patch(
-        "ChemInformant.cheminfo_api.api_helpers.get_batch_synonyms"
-    )
-    mock_get_batch_desc = mocker.patch(
[…]
