### 2025-07-01 ad0eb1d
Delete src/ChemInformant/__init__.py
