### 2025-07-01 c2e5a6d
Delete src/ChemInformant/api_helpers.py
