+++ src/ChemInformant/api_helpers.py
@@ -1,459 +1,0 @@
-# src/ChemInformant/api_helpers.py
-import requests
-import requests_cache  # Import requests_cache here
-import sys
-
-# import time # Removed unused import
-from urllib.parse import quote
-import xml.etree.ElementTree as ET
-from typing import Optional, Tuple, List, Dict, Any
-
-# --- Constants ---
-PUBCHEM_API_BASE = "https://pubchem.ncbi.nlm.nih.gov/rest/pug"
-PUG_VIEW_BASE = "https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data"
-REQUEST_TIMEOUT = 15
-
-# --- Caching Setup (Moved from __init__.py) ---
-DEFAULT_CACHE_NAME = "pubchem_cache"
-DEFAULT_CACHE_BACKEND = "sqlite"
-DEFAULT_CACHE_EXPIRE_AFTER = 60 * 60 * 24 * 7  # 7 days in seconds
-
-_session = None  # Module-level session variable within api_helpers
-
-
-def setup_cache(
-    cache_name=DEFAULT_CACHE_NAME,
-    backend=DEFAULT_CACHE_BACKEND,
-    expire_after=DEFAULT_CACHE_EXPIRE_AFTER,
-    **kwargs,
-):
-    """
-    Configures the cache used for PubChem API requests.
-
-    Call this function *before* making any ChemInformant API calls if you
-    want to use non-default cache settings. This function should be imported
-    from the ChemInformant package level (e.g., `ChemInformant.setup_cache(...)`).
-
-    Args:
-        cache_name (str): Name of the cache file/db.
-        backend (str): Cache backend ('sqlite', 'memory', etc.).
-        expire_after (int | float | None | timedelta): Expiration time in seconds.
-        **kwargs: Additional backend arguments for requests_cache.CachedSession.
-    """
-    global _session
-    # Create a new session each time setup_cache is called
-    _session = requests_cache.CachedSession(
-        cache_name=cache_name,
-        backend=backend,
-        expire_after=expire_after,
-        allowable_codes=[200, 404],
-        match_headers=False,
-        **kwargs,
-    )
-    # print(f"ChemInformant cache setup: name='{cache_name}', backend='{backend}', expire_after={expire_after}s") # Optional print
-
-
-def get_session():
-    """Internal function to get the current session (cached or default)."""
-    global _session  # noqa: F824
-    if _session is None:
-        # Initialize with defaults if setup_cache wasn't called explicitly
-        # print("ChemInformant: Initializing default cache...") # Optional print
-        setup_cache()  # Call setup_cache with defaults
-        if _session is None:  # Safeguard
-            raise RuntimeError("Failed to initialize requests session.")
-    return _session
-
-
-# --- Core Fetch Function (uses session defined above) ---
-def _fetch_data(
-    url: str, identifier_type: str = "CID", identifier: Any = None
-) -> Optional[Any]:
-    """Fetches data using the configured session. Returns parsed content or None."""
-    session = get_session()  # Calls internal get_session
-    try:
-        response = session.get(url, timeout=REQUEST_TIMEOUT)
-        # was_cached = getattr(response, 'from_cache', False) # Debugging
-        # print(f"DEBUG: Cache hit={was_cached} for URL: {url}", file=sys.stderr)
-
-        if response.status_code == 404:
-            return None
-        response.raise_for_status()  # Raise for other 4xx/5xx errors
-
-        content_type = response.headers.get("Content-Type", "").lower()
-        if "application/json" in content_type:
-            return response.json()
-        elif "application/xml" in content_type or "text/xml" in content_type:
-            # Decode XML content assuming UTF-8, handle potential errors
-            try:
[…]
