### 2025-07-01 b28661c
Delete src/ChemInformant/models.py
