### 2025-07-01 4e0603a
Delete src/ChemInformant directory
