+++ src/ChemInformant/cheminfo_api.py
@@ -1,485 +1,0 @@
-# src/ChemInformant/cheminfo_api.py
-"""
-Provides simple, cached, validated functions for retrieving chemical information
-from PubChem, including ambiguity handling and batch retrieval.
-"""
-import sys
-from typing import Optional, List, Union, Dict, Any, Tuple
-import time  # For potential sleeps in batch
-import io # Standard library, can stay at top
-import requests # Explicitly import requests
-
-# Use relative imports within the package
-from . import api_helpers
-from .models import CompoundData, NotFoundError, AmbiguousIdentifierError
-
-# --- Imports for image display ---
-import io
-
-
-# --- Internal Helper ---
-def _resolve_identifier(identifier: Union[str, int]) -> Union[int, List[int]]:
-    """
-    Resolves an identifier (name or CID) to a PubChem CID or list of CIDs.
-
-    Raises:
-        NotFoundError: If the identifier cannot be found.
-        AmbiguousIdentifierError: If a name maps to multiple CIDs (contains the list).
-        ValueError: If an invalid integer CID (<1) is provided.
-        TypeError: If the input type is incorrect.
-    """
-    if isinstance(identifier, int):
-        if identifier > 0:
-            return identifier
-        else:
-            raise ValueError(f"Invalid CID value provided: {identifier}")
-    elif isinstance(identifier, str):
-        cids = api_helpers.get_cids_by_name(identifier)
-        if cids is None or len(cids) == 0:
-            raise NotFoundError(identifier=identifier)
-        elif len(cids) == 1:
-            return cids[0]  # Unambiguous name
-        else:
-            raise AmbiguousIdentifierError(identifier=identifier, cids=cids)
-    else:
-        raise TypeError(
-            f"Input must be a compound name (str) or CID (int), got {type(identifier)}"
-        )
-
-
-# --- Primary Single Compound Function ---
-def info(name_or_cid: Union[str, int]) -> CompoundData:
-    """
-    Retrieves comprehensive, validated details for a single compound.
-    Attempts to return partial data even if some underlying fetches fail.
-
-    Args:
-        name_or_cid: The compound name (str) or PubChem CID (int).
-
-    Returns:
-        CompoundData: A Pydantic model instance with the compound information.
-
-    Raises:
-        NotFoundError: If the compound identifier cannot be resolved initially.
-        AmbiguousIdentifierError: If the name maps to multiple CIDs initially.
-        ValueError: If an invalid CID (<1) is provided initially.
-        TypeError: If the input type is incorrect initially.
-        Exception: Potentially Pydantic validation errors if data is severely malformed.
-    """
-    resolved_id = _resolve_identifier(name_or_cid)  # Raises initial errors if needed
-
-    if not isinstance(resolved_id, int):
-        raise TypeError(f"Internal error: Expected single CID but got {resolved_id}")
-
-    cid_val = resolved_id
-
-    # --- Fetch components individually, handling potential errors for each ---
-    cas_val, unii_val = None, None
-    props = None
-    description_val = None
-    synonyms_val = None
-
-    try:
-        cas_val, unii_val = api_helpers.get_cas_unii(cid_val)
-    except Exception as e:
-        print(
-            f"Warning: Failed to get CAS/UNII for CID {cid_val}: {type(e).__name__}",
-            file=sys.stderr,
-        )
-
-    try:
-        props = api_helpers.get_additional_properties(cid_val)
-    except Exception as e:
-        print(
-            f"Warning: Failed to get additional properties for CID {cid_val}: {type(e).__name__}",
-            file=sys.stderr,
-        )
-
-    try:
-        description_val = api_helpers.get_compound_description(cid_val)
-    except Exception as e:
-        print(
[…]
