### 2025-07-01 57afe4b
Delete examples directory
