+++ benchmark.py
@@ -1,0 +1,135 @@
+import time
+import ChemInformant as ci
+import pandas as pd
+import pubchempy as pcp
+
+def benchmark_pubchempy(identifiers, properties_to_map):
+    """
+    Benchmark using PubChemPy with individual queries as a baseline.
+    """
+    print("\n[3] Benchmarking PubChemPy (individual queries)...")
+    start_time = time.perf_counter()
+    
+    results = []
+    
+    for identifier in identifiers:
+        record = {'input_identifier': identifier, 'status': 'UnknownError'}
+        try:
+            compounds = pcp.get_compounds(identifier, 'name')
+            if not compounds:
+                record['status'] = 'NotFoundError'
+                results.append(record)
+                continue
+
+            c = compounds[0]
+            record.update({'status': 'OK', 'cid': c.cid})
+            
+            if 'molecular_weight' in properties_to_map:
+                record['molecular_weight'] = getattr(c, 'molecular_weight', None)
+            if 'molecular_formula' in properties_to_map:
+                record['molecular_formula'] = getattr(c, 'molecular_formula', None)
+            if 'xlogp' in properties_to_map:
+                record['xlogp'] = getattr(c, 'xlogp', None)
+            if 'canonical_smiles' in properties_to_map:
+                record['canonical_smiles'] = getattr(c, 'canonical_smiles', None)
+
+            if 'cas' in properties_to_map:
+                cas_list = [syn for syn in c.synonyms if syn.isdigit() and len(syn.split('-')) == 3]
+                record['cas'] = cas_list[0] if cas_list else None
+            
+            if 'synonyms' in properties_to_map:
+                record['synonyms'] = getattr(c, 'synonyms', [])
+            
+            results.append(record)
+
+        except Exception as e:
+            record['status'] = type(e).__name__
+            results.append(record)
+
+    duration = time.perf_counter() - start_time
+    print(f"    > Done! Took {duration:.4f} seconds.")
+    return duration, pd.DataFrame(results)
+
+def run_benchmark(identifiers, properties):
+    """
+    Run benchmark to measure uncached, cached, and PubChemPy comparison performance.
+    """
+    print("--- ChemInformant Performance Benchmark ---")
+    print(f"Fetching {len(properties)} properties for {len(identifiers)} compounds...")
+
+    ci.setup_cache(backend='memory')
+    print("In-memory cache set up. Benchmark starting...")
+
+    # 1. ChemInformant (uncached)
+    print("\n[1] ChemInformant (uncached) batch call...")
+    start_time_uncached = time.perf_counter()
+    df_ci_uncached = ci.get_properties(identifiers, properties)
+    duration_uncached = time.perf_counter() - start_time_uncached
+    print(f"    > Done! Took {duration_uncached:.4f} seconds.")
+
+    ok_count = (df_ci_uncached['status'] == 'OK').sum()
+    error_count = len(identifiers) - ok_count
+    print(f"    > Results: {ok_count} success, {error_count} failures.")
+    if ok_count == 0:
+        print("    > Warning: Failed to fetch any data. Check network connection.")
+        return
+
+    # 2. ChemInformant (cached)
+    print("\n[2] ChemInformant (cached) second call...")
+    start_time_cached = time.perf_counter()
+    df_ci_cached = ci.get_properties(identifiers, properties)
+    duration_cached = time.perf_counter() - start_time_cached
+    print(f"    > Done! Took {duration_cached:.4f} seconds.")
+
+    # 3. PubChemPy (as baseline)
[…]
