+++ README.md
@@ -11,302 +11,4 @@ ChemInformant <img src="https://github.com/HzaCode/ChemInformant/blob/main/image
 
 
 
-**ChemInformant** is a Python library designed for streamlined access to chemical data from the [PubChem](https://pubchem.ncbi.nlm.nih.gov/) database. It simplifies retrieving compound information by providing:
-
-* **Automatic Caching:** Uses `requests-cache` to avoid redundant API calls and speed up repeated queries.
-* **Data Validation:** Employs `pydantic` models (`CompoundData`) to ensure retrieved data is structured and typed correctly.
-* **Robust Error Handling:** Gracefully handles cases where compounds are not found (`NotFoundError`) or identifiers are ambiguous (`AmbiguousIdentifierError`).
-* **Convenience Functions:** Offers simple functions (`cid`, `cas`, `form`, `wgt`, etc.) to fetch specific properties.
-* **Batch Processing:** Efficiently retrieves data for multiple compounds in a single call (`get_multiple_compounds`).
-* **Partial Data Retrieval:** Attempts to return as much data as possible even if some underlying API calls fail for a specific compound.
-* **Image Display:** Directly displays 2D chemical structures within your environment (e.g., Jupyter, VS Code interactive) using `fig()`.
----
-🎉 **New Feature Launch!** 🎉
-
-**Visualize Chemical Structures!**
-
-ChemInformant now empowers you to generate and display 2D chemical structures directly within your Python environment. Simply use the new `ci.fig()` function!
-
-This makes visualizing compounds **quick and easy**. For detailed examples, see the [Basic Usage](#basic-usage) and [API Overview](#api-overview) sections.
-
----
-## Key Features
-
-* Look up compounds by name or PubChem CID.
-* Retrieve comprehensive compound details (`info`) or specific properties (`cas`, `unii`, `molecular_formula`, `molecular_weight`, `canonical_smiles`, `iupac_name`, `description`, `synonyms`).
-* Automatic caching of API responses to disk (default: `pubchem_cache.sqlite`) or memory.
-* Configurable cache expiration time (default: 7 days).
-* Clear distinction between "not found" and "ambiguous identifier" errors.
-* Efficient batch retrieval for lists of identifiers.
-* Validated and typed data output via the `CompoundData` model.
-* Computed `pubchem_url` property on the `CompoundData` model.
-* Inline display of 2D compound structures (`ci.fig()`).
-
-## Installation
-
-Install the latest stable version from PyPI:
-
-```bash
-pip install ChemInformant
-```
-*(Requires Python >= 3.8)*
-
-Or, install directly from the GitHub repository (for the latest development version):
-
-```bash
-pip install git+https://github.com/HzaCode/ChemInformant.git
-
-```
-
-For editable installs :
-```bash
-# Clone the repository first if you haven't
- git clone https://github.com/HzaCode/ChemInformant.git
- cd ChemInformant
-
-pip install -e .
-```
-
-To use the image display feature (`ci.fig()`), you will also need `matplotlib` and `Pillow`:
-
-```bash
-pip install matplotlib Pillow
-```
-
-## Basic Usage
-
-The recommended way to use the library is by importing it with the alias `ci`.
-
-```python
-import ChemInformant as ci
-import os
-import sys # Needed for stderr printing in case of warnings
-
-# --- Optional: Configure Cache (before first API call) ---
-# Default: 'pubchem_cache.sqlite' in current dir, 7 days expiry
-# Example: Use in-memory cache expiring after 1 hour
-# ci.setup_cache(backend='memory', expire_after=3600)
-
[…]
