### 2025-07-04 d61c036
Delete paper directory
