+++ paper/paper.md
@@ -1,117 +1,0 @@
----
-title: 'ChemInformant: A Modern, Lightweight Python Client for PubChem with Robust Caching and Validation'
-tags:
-  - python
-  - pubchem
-  - chemistry
-  - api
-  - cache
-  - pydantic
-  - cheminformatics
-  - research software
-  - data validation
-authors:
-  - name: Zhiang He
-    affiliation: 1
-affiliations:
-  - name: Independent researcher
-    index: 1
-date: "2025-04-15"
-bibliography: paper.bib 
-repository: https://github.com/HzaCode/ChemInformant
-version: 1.2.1 
-license: MIT
----
-
-
-## Summary
-
-ChemInformant provides a **lightweight Pythonic interface** characterized by its ease of use for accessing chemical compound data from the PubChem database [@PubChem; @Kim2016NAR]. Designed to abstract the complexities of the PubChem API (`import ChemInformant as ci`), it facilitates streamlined data retrieval for researchers. Key distinguishing features include **built-in automatic caching** via `requests-cache` [@RequestsCache], which enhances performance and contributes to reproducibility; **structured data validation** using Pydantic [@Pydantic] models (`CompoundData`), promoting data quality; **explicit handling of chemical name ambiguity**, enhancing accuracy; and **optimized batch retrieval**. Furthermore, ChemInformant is built upon a minimal set of **modern, actively maintained core dependencies** (`requests` [@Requests], `requests-cache`, `Pydantic`), which contributes to its long-term stability and compatibility. This design addresses certain limitations associated with direct API utilization and offers enhancements over existing tools that may be built upon less contemporary or less actively maintained software foundations. ChemInformant is designed as a reliable and efficient tool for routine chemical information retrieval.
-
-## Statement of Need
-
-Programmatic access to PubChem [@PubChem] is integral for numerous chemistry-related research tasks; however, direct utilization of the PUG REST API [@Kim2018JCheminform; @Kim2018PUGREST] can entail considerable overhead in request handling, response parsing, error management, and data validation. This inherent complexity can impede research progress and potentially introduce errors.
-
-While existing libraries, such as PubChemPy [@PubChemPy], provide an interface to PubChem, `ChemInformant` was developed to address specific limitations and to offer a contemporary, robust, and efficient alternative tailored to common research workflows by focusing on the following aspects:
-
-1.  **Dependency Modernity and Maintenance:** `ChemInformant` is designed to rely on a small set of ubiquitous, actively developed, and well-maintained libraries (`requests` [@Requests], `requests-cache` [@RequestsCache], `Pydantic` [@Pydantic]). This approach supports long-term stability, security, and compatibility with evolving Python [@Python] environments.
-2.  **Integrated Automatic Caching:** `ChemInformant` integrates `requests-cache` **by default**. This feature provides transparent, persistent caching, which substantially improves performance for repeated queries and batch operations, enhances operational stability during network variability, and contributes to the reproducibility of analyses. The caching is also readily configurable (`ci.setup_cache()`).
-3.  **Rigorous Data Validation and Structure:** `ChemInformant` employs Pydantic to define a clear `CompoundData` model, ensuring that retrieved data are validated at runtime, data types are enforced, and users interact with consistently structured objects.
-4.  **Explicit Ambiguity Handling:** `ChemInformant` explicitly addresses chemical name ambiguity by raising a distinct `AmbiguousIdentifierError` containing all potential CIDs, thereby preventing silent failures or arbitrary selections.
-5.  **Simplicity and Ease of Use:** The API is designed to be Pythonic and intuitive (e.g., `ci.info()`, `ci.cas()`), reducing the adoption barrier for researchers seeking to integrate PubChem data access into Python scripts or Jupyter notebooks [@Jupyter].
-
[…]
