+++ docs/_sources/index.rst
@@ -1,25 +1,0 @@
-#########################################
-Welcome to ChemInformant's Documentation!
-#########################################
-
-
-.. toctree::
-   :maxdepth: 1
-   :caption: User Guide
-
-   usage
-
-.. toctree::
-   :caption: API Reference
-   :maxdepth: 2
-
-   API Overview <api/index>
-   ChemInfo API <api/ChemInformant.cheminfo_api>
-   Data Models <api/ChemInformant.models>
-   API Helpers <api/ChemInformant.api_helpers>
-
-.. toctree::
-   :caption: Development
-   :maxdepth: 1
-
-   changelog
