+++ docs/_sources/api/ChemInformant.api_helpers.rst
@@ -1,7 +1,0 @@
-ChemInformant.api\_helpers module
-=================================
-
-.. automodule:: ChemInformant.api_helpers
-   :members:
-   :show-inheritance:
-   :undoc-members:
