+++ README.md
@@ -1,8 +1,6 @@
-# ChemInformant <img src="https://github.com/HzaCode/ChemInformant/blob/joss-review-response/images/logo.jpg?raw=true" align="right" width="120px" />
-
+# ChemInformant <img src="https://raw.githubusercontent.com/HzaCode/ChemInformant/joss-review-response/images/logo.png" align="right" width="120px" />
 
 [![Status: Revisions in Progress](https://img.shields.io/badge/Status-Revisions%20in%20Progress-orange.svg?style=flat-square)](#)
-
 
 **ChemInformant** is a modern, workflow-centric **Python client** for the [PubChem](https://pubchem.ncbi.nlm.nih.gov/) database. It's designed to feel native and intuitive for researchers and data scientists working within the **Python data science ecosystem (e.g., Pandas, Jupyter)**.
 
@@ -12,21 +10,12 @@ It's built to be robust by default, with built-in request caching, smart rate li
 
 ### Key Features
 
-*   **Workflow-First Design (Native Pandas Output):** The core API (`get_properties`) directly returns a clean **Pandas DataFrame**, eliminating boilerplate code and allowing for immediate integration with other data science libraries.
-
-*   **Robust by Default (Caching, Rate-Limiting & Retries):**
-    *   **Transparent Caching:** Built-in, persistent caching significantly accelerates repetitive queries.
-    *   **Smart Rate-Limiting & Retries:** Automatically adheres to PubChem's usage policies and gracefully handles temporary server issues with an exponential backoff mechanism.
-
-*   **Intuitive Layered API:**
-    *   A **convenience layer** of `get_<property>()` functions for quick, simple lookups.
-    *   A **powerful engine layer** (`get_properties`) for complex, performance-critical batch operations.
-
-*   **Automated Pagination:** Transparently handles PubChem's `ListKey`-based pagination, effortlessly retrieving complete result sets from large queries without manual intervention.
-
-*   **Type-Safe and Validated:** Employs Pydantic v2 models (`Compound`) to ensure the data you receive is structured and type-checked, preventing common errors downstream.
-
-*   **Modern Tech Stack & Maintenance:** Built on a minimal set of **actively maintained libraries** (like `requests-cache` and `Pydantic v2`), ensuring long-term stability and compatibility, in contrast to older, less frequently updated tools.
+*   **Native Pandas Output:** Core API returns a clean, ready-to-analyze **Pandas DataFrame**.
+*   **Robust by Default:** Built-in caching, rate-limiting, and retries are enabled with zero configuration.
+*   **Intuitive Layered API:** Offers both simple `get_<property>()` functions for quick lookups and a powerful `get_properties` engine for batch operations.
+*   **Automated Pagination:** Transparently handles PubChem's `ListKey`-based pagination for large queries.
+*   **Type-Safe & Validated:** Employs Pydantic v2 models to ensure the data you receive is structured and type-checked.
+*   **Modern & Maintained:** Built on an actively maintained tech stack for long-term stability.
 
 ### Installation
 
@@ -41,16 +30,17 @@ To include plotting capabilities for use with the tutorial, install the `[plot]`
 pip install "ChemInformant[plot]"
 ```
 
+---
+
 ### Quick Start
 
 Retrieve multiple properties for multiple compounds, directly into a Pandas DataFrame, in a single function call:
 
 ```python
 import ChemInformant as ci
-import pandas as pd
 
 # 1. Define your identifiers
-identifiers = ["aspirin", "caffeine", "paracetamol", "NonExistentCompound"]
+identifiers = ["aspirin", "caffeine", 1983] # 1983 is paracetamol's CID
 
 # 2. Specify the properties you need
 properties = ["molecular_weight", "xlogp", "cas"]
[…]
