+++ pyproject.toml
@@ -4,14 +4,14 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ChemInformant"
-version = "1.2.2"
+version = "1.3.0"
 authors = [
-  { name="Zhiang He", email="ang@hezhiang.com" },
+  { name = "Zhiang He", email = "ang@hezhiang.com" },
 ]
 description = "A Python library for streamlined access to PubChem chemical data with caching, validation, ambiguity handling, and batch retrieval."
-readme = "README.md"
+readme = { file = "README.md", content-type = "text/markdown" }
 requires-python = ">=3.8"
-license = { file="LICENSE" }
+license = { file = "LICENSE" }
 classifiers = [
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.8",
@@ -23,25 +23,50 @@ classifiers = [
     "Operating System :: OS Independent",
     "Topic :: Scientific/Engineering :: Chemistry",
 ]
-keywords = ["chemistry", "cheminformatics", "pubchem", "api", "compound", "drug", "cache", "pydantic", "batch"]
+keywords = [
+    "chemistry",
+    "cheminformatics",
+    "pubchem",
+    "api",
+    "compound",
+    "drug",
+    "cache",
+    "pydantic",
+    "batch",
+    "smiles"
+]
+
 dependencies = [
     "requests>=2.25.0",
-    "requests-cache>=1.1.0",
+    "requests-cache>=1.2.0",      
     "pydantic>=2.0",
     "pandas>=1.5.0"
 ]
 
 [project.urls]
-Homepage = "https://github.com/HzaCode/ChemInformant"
-Issues = "https://github.com/HzaCode/ChemInformant/issues"
+Homepage       = "https://github.com/HzaCode/ChemInformant"
+Documentation  = "https://github.com/HzaCode/ChemInformant#readme"
+Issues         = "https://github.com/HzaCode/ChemInformant/issues"
 
 [project.optional-dependencies]
 plot = [
     "matplotlib>=3.1",
     "Pillow>=9.0"
 ]
-dev = ["pytest>=7.0", "pytest-mock>=3.5", "flake8>=6.0", "pytest-cov>=2.12"]
-all = ["ChemInformant[plot,dev]"]
+dev = [
+    "pytest>=7.0",
+    "pytest-mock>=3.5",
+    "flake8>=6.0",
+    "pytest-cov>=2.12"
+]
+all = [
+    "matplotlib>=3.1",
+    "Pillow>=9.0",
+    "pytest>=7.0",
+    "pytest-mock>=3.5",
+    "flake8>=6.0",
+    "pytest-cov>=2.12"
+]
 
 [tool.setuptools.packages.find]
 where = ["src"]
