+++ README.md
@@ -87,7 +87,7 @@ print(df)
 
 This quick start only scratches the surface. For a deep dive, please see our detailed guides:
 
-*   **➡️ Interactive User Manual:** Our [**Jupyter Notebook Tutorial**](./ChemInformant_User_Manual.ipynb) provides a complete, end-to-end walkthrough. **This is the best place to start.**
+*   **➡️ Interactive User Manual:** Our [**Jupyter Notebook Tutorial**](examples/ChemInformant_User_Manual_v1.0.ipynb)provides a complete, end-to-end walkthrough. **This is the best place to start.**
 *   **➡️ Performance Benchmarks:** You can review and run our [**Benchmark Script**](./benchmark.py) to see the performance advantages of batching and caching.
 
 ---
