+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ChemInformant"
-version = "1.3.0"
+version = "2.0.0"
 authors = [
   { name = "Zhiang He", email = "ang@hezhiang.com" },
 ]
@@ -51,7 +51,9 @@ Issues         = "https://github.com/HzaCode/ChemInformant/issues"
 [project.optional-dependencies]
 plot = [
     "matplotlib>=3.1",
-    "Pillow>=9.0"
+    "Pillow>=9.0",
+    "seaborn>=0.12", 
+    "scikit-learn>=1.1"
 ]
 dev = [
     "pytest>=7.0",
