+++ tests/test_api.py
@@ -1,1 +1,178 @@
+from __future__ import annotations
 
+import json, re, sys, types, pytest
+import ChemInformant as ci
+from ChemInformant import api_helpers as ah
+from ChemInformant import cheminfo_api as capi
+from ChemInformant import models
+
+
+# ------------------------------------------------------------------#
+# 0. Use in-memory HTTP cache
+@pytest.fixture(autouse=True)
+def _mem_cache():
+    ci.setup_cache(backend="memory")
+
+
+# ------------------------------------------------------------------#
+# 1. Fake CachedSession (implements cache.disabled())
+class _DummyCtx:
+    def __enter__(self): ...
+    def __exit__(self, exc_type, exc, tb): ...
+
+
+class _Cache:
+    def create_key(self, _): return "K"
+    def delete(self, _):     ...
+    def disabled(self):      return _DummyCtx()
+
+
+class _Session:
+    def __init__(self):
+        self.cache = _Cache()
+
+    def get(self, url, timeout=None):
+        return _fake_execute_fetch(url)
+
+
+# ------------------------------------------------------------------#
+# 2. Fake HTTP response
+class _Resp:
+    def __init__(self, status=200, data: dict | None = None,
+                 from_cache=False,
+                 ctype="application/json"):
+        self.status_code = status
+        self.from_cache = from_cache
+        self.headers = {"Content-Type": ctype}
+        self.request = object()
+        self._data = data or {}
+
+    def json(self): return self._data
+    @property
+    def text(self): return json.dumps(self._data)
+
+
+# ------------------------------------------------------------------#
+# 3. Stub all network traffic
+def _fake_execute_fetch(url: str) -> _Resp:                       # noqa: C901
+    """Return canned PubChem-like JSON for the URLs used in tests."""
+    # name/smiles → CID
+    m = re.search(r"/compound/(name|smiles)/([^/]+)/cids", url)
+    if m:
+        token = re.sub(
+            r"%..", lambda m: bytes.fromhex(m.group()[1:]).decode(), m.group(2)
+        )
+        mapping = {"aspirin": [2244], "ambiguous": [1, 2], "C1CC1": [999]}
+        return _Resp(data={"IdentifierList": {"CID": mapping.get(token, [])}})
+    # batch properties with pagination
+    if "/property/" in url:
+        if "/listkey/" in url:  # second page
+            return _Resp(data={"PropertyTable": {"Properties": [
+                {"CID": 222, "Foo": "b"}
+            ]}})
+        first_page = {"PropertyTable": {"Properties": [
+            {"CID": 111, "Foo": "a"},
+            {"CID": 2244, "MolecularWeight": 180.16,
+             "MolecularFormula": "C9H8O4",
+             "ConnectivitySMILES": "CC(=O)Oc1ccccc1C(=O)O"},
+            {"CID": 999, "MolecularWeight": 46.07,
+             "CanonicalSMILES": "C1CC1",
+             "MolecularFormula": "C3H6"},
+        ]}, "ListKey": "NEXT"}
+        return _Resp(data=first_page)
+    # CAS lookup
+    if "/pug_view/data/compound/" in url:
+        return _Resp(data={"Record": {"Section": [{
+            "TOCHeading": "Names and Identifiers",
+            "Section": [{
+                "TOCHeading": "Other Identifiers",
+                "Section": [{
+                    "TOCHeading": "CAS",
+                    "Information": [{
+                        "Value": {"StringWithMarkup": [{"String": "50-78-2"}]}
+                    }]
+                }]
[…]
