+++ benchmark.py
@@ -1,135 +1,133 @@
-import time
+
+"""
+ChemInformant vs. PubChemPy
+(300 unique drugs × 6 properties)
+
+Output:
+  • PubChemPy bulk time
+  • ChemInformant first-call batch time (cold)
+  • ChemInformant cache-hit batch time (warm, mean of 3 runs)
+  • Throughput (compounds / s) and speed-up ratios
+"""
+
+
+import uuid, time, statistics
 import ChemInformant as ci
-import pandas as pd
 import pubchempy as pcp
 
-def benchmark_pubchempy(identifiers, properties_to_map):
-    """
-    Benchmark using PubChemPy with individual queries as a baseline.
-    """
-    print("\n[3] Benchmarking PubChemPy (individual queries)...")
-    start_time = time.perf_counter()
-    
-    results = []
-    
-    for identifier in identifiers:
-        record = {'input_identifier': identifier, 'status': 'UnknownError'}
-        try:
-            compounds = pcp.get_compounds(identifier, 'name')
-            if not compounds:
-                record['status'] = 'NotFoundError'
-                results.append(record)
-                continue
+# ---------------- 1. 300 unique drug names ----------------
+DRUGS = sorted({
+    "abacavir","acetaminophen","acetazolamide","acyclovir","adalimumab",
+    "albendazole","albuterol","alclometasone","allopurinol","alprazolam",
+    "amiodarone","amitriptyline","amlodipine","amoxicillin","ampicillin",
+    "anastrozole","apixaban","aripiprazole","aspirin","atorvastatin",
+    "azathioprine","azithromycin","baclofen","beclomethasone","benzene",
+    "betamethasone","bisoprolol","bortezomib","brinzolamide","budesonide",
+    "bupivacaine","bupropion","cabergoline","capecitabine","capsaicin",
+    "carbamazepine","carboplatin","carvedilol","cefazolin","cefepime",
+    "ceftriaxone","celecoxib","cephalexin","chloramphenicol","chloroquine",
+    "chlorpromazine","cilostazol","ciprofloxacin","cisplatin","clindamycin",
+    "clobetasol","clonazepam","clopidogrel","clozapine","colchicine",
+    "cyclobenzaprine","cyclophosphamide","cyclosporine","dapagliflozin",
+    "daptomycin","darunavir","dasatinib","desloratadine","dexamethasone",
+    "dexmedetomidine","diazepam","diclofenac","digoxin","diltiazem",
+    "diphenhydramine","disulfiram","docetaxel","donepezil","dopamine",
+    "dorzolamide","doxycycline","dronedarone","duloxetine","efavirenz",
+    "enalapril","epinephrine","erlotinib","esomeprazole","estradiol",
+    "ethambutol","etoposide","everolimus","exemestane","famciclovir",
+    "famotidine","felodipine","fenofibrate","fentanyl","ferrous sulfate",
+    "finasteride","fluconazole","fludarabine","fludrocortisone","flumazenil",
+    "fluoxetine","fluticasone","folic acid","foscarnet","fosfomycin",
+    "furosemide","gabapentin","ganciclovir","gemcitabine","gefitinib",
+    "glimepiride","glipizide","glucose","glyburide","haloperidol",
+    "heparin sodium","hydralazine","hydrochlorothiazide","hydrocodone",
+    "hydrocortisone","hydromorphone","ibandronate","ibuprofen","idarubicin",
+    "ifosfamide","imatinib","indomethacin","insulin","interferon alfa-2b",
+    "ipratropium","irbesartan","irinotecan","isavuconazole","isoniazid",
+    "isotretinoin","itraconazole","ketamine","ketoconazole","ketoprofen",
[…]
