### 2025-07-17 63f478c
Update api_helpers.py
