### 2025-07-17 587709a
Update cheminfo_api.py
