+++ src/ChemInformant/cheminfo_api.py
@@ -1,1 +1,166 @@
 
+from __future__ import annotations
+
+import re
+from typing import Iterable, List, Union, Dict, Any
+
+import pandas as pd
+
+from . import api_helpers
+from .models import Compound, NotFoundError, AmbiguousIdentifierError
+
+# SMILES pattern tokens
+_SMILES_TOKENS = re.compile(r"[=#\[\]\(\)]|\d|Br|Cl|Si", re.I)
+
+def _looks_like_smiles(s: str) -> bool:
+    return bool(_SMILES_TOKENS.search(s))
+
+def _resolve_to_single_cid(identifier: Union[str, int]) -> int:
+    if isinstance(identifier, int):
+        if identifier > 0:
+            return identifier
+        raise ValueError(f"Invalid CID: {identifier}")
+    if isinstance(identifier, str) and identifier.isdigit():
+        return int(identifier)
+
+    if isinstance(identifier, str):
+        cids = api_helpers.get_cids_by_name(identifier)
+        if cids:
+            if len(cids) == 1:
+                return cids[0]
+            raise AmbiguousIdentifierError(identifier, cids)
+
+    if isinstance(identifier, str) and _looks_like_smiles(identifier):
+        cids = api_helpers.get_cids_by_smiles(identifier)
+        if cids:
+            if len(cids) == 1:
+                return cids[0]
+            raise AmbiguousIdentifierError(identifier, cids)
+
+    raise NotFoundError(identifier)
+
+PROPERTY_ALIASES: Dict[str, List[str]] = {
+    "molecular_weight":   ["MolecularWeight"],
+    "molecular_formula":  ["MolecularFormula"],
+    "canonical_smiles":   ["CanonicalSMILES", "ConnectivitySMILES"],
+    "isomeric_smiles":    ["IsomericSMILES", "SMILES"],
+    "iupac_name":         ["IUPACName"],
+    "xlogp":              ["XLogP"],
+}
+_SPECIAL_PROPS = {"cas", "synonyms"}
+
+def get_properties(
+    identifiers: Iterable[Union[str, int]],
+    properties: Iterable[str],
+) -> pd.DataFrame:
+    identifiers = list(identifiers)
+    properties  = list(properties)
+
+    if not identifiers or not properties:
+        return pd.DataFrame()
+
+    regular     = [p for p in properties if p in PROPERTY_ALIASES]
+    specials    = [p for p in properties if p in _SPECIAL_PROPS]
+    unsupported = [p for p in properties if p not in regular + specials]
+    if unsupported:
+        raise ValueError(f"Unsupported properties: {unsupported}")
+
+    meta: Dict[Any, Dict[str, Any]] = {}
+    cids_needed: set[int] = set()
+    for ident in identifiers:
+        try:
+            cid = _resolve_to_single_cid(ident)
+            meta[ident] = {"status": "OK", "cid": cid}
+            cids_needed.add(cid)
+        except Exception as exc:
+            meta[ident] = {"status": type(exc).__name__, "cid": None}
+
+    fetched_regular: Dict[int, Dict[str, Any]] = {}
+    if cids_needed and regular:
+        tags: List[str] = []
+        for p in regular:
+            tags.extend(PROPERTY_ALIASES[p])
+        tags = list(dict.fromkeys(tags))
+        fetched_regular = api_helpers.get_batch_properties(list(cids_needed), tags)
+
+    fetched_special = {sp: {cid: None for cid in cids_needed} for sp in specials}
+    for cid in cids_needed:
+        if "cas" in specials:
+            fetched_special["cas"][cid] = api_helpers.get_cas_for_cid(cid)
+        if "synonyms" in specials:
[…]
