### 2025-07-17 2338ec4
Update CITATION.cff
