### 2025-07-17 25b9c5b
Update paper.md
