+++ paper/paper.md
@@ -1,1 +1,111 @@
+---
+title: 'ChemInformant: A Robust and Workflow-Centric Python Client for High-Throughput PubChem Access'
+tags:
+  - python
+  - pubchem
+  - chemistry
+  - api
+  - pandas
+  - cache
+  - pydantic
+  - cheminformatics
+  - data science
+  - high-throughput
+authors:
+  - name: Zhiang He
+    affiliation: 1
+    
+affiliations:
+  - name: Independent Researcher
+    index: 1
+date: "2025-04-15"
+bibliography: paper.bib
+repository: https://github.com/HzaCode/ChemInformant
+version: 2.0.0
+license: MIT
+---
+
+## Summary
+
+ChemInformant is a modern, workflow-centric Python client designed for high-throughput and automated PubChem data tasks. Its architecture is engineered to support the direct conversion of various chemical identifiers, including mixed-type lists, into "analysis-ready" Pandas DataFrames [@Pandas], in order to streamline the entire workflow from data retrieval to scientific analysis. As a natively integrated solution, the tool enables zero-configuration robustness by default, incorporating mechanisms such as persistent caching, automatic rate-limiting with exponential backoff, and runtime type validation with Pydantic [@Pydantic]. In benchmark tests, ChemInformant's initial query was 4.6 times faster than the batch interface of a mainstream library; with a warm cache, the performance advantage increased to 48-fold, achieving response times suitable for interactive analysis tools. By systematically addressing key limitations of existing tools in network reliability, batch processing flexibility, and maintainability, ChemInformant is positioned to serve as an efficient and reliable data infrastructure component for the Python cheminformatics ecosystem.
+
+## Statement of Need
+
+Programmatic access to PubChem [@PubChem] has become a critical infrastructure component in modern chemical and life sciences research. As research scales and automated workflows become more prevalent, researchers face three core challenges with existing client tools: the reliability of network access, the flexibility and fault tolerance of batch processing, and the sustainability risks associated with long-term library maintenance.
+
+First, network reliability is a primary concern. PubChem's API service enforces strict dynamic rate limits (typically ≤5 requests per second) and returns `HTTP 503 (Server Busy)` errors under high load [@PubChemUsagePolicy]. Mainstream clients like PubChemPy [@PubChemPy] generally lack built-in automatic throttling and exponential backoff retry mechanisms, which makes automated scripts brittle when faced with real-world network fluctuations. Community feedback widely reports such stability issues, with common workarounds still relying on manual insertion of delays. Furthermore, the absence of a persistent caching mechanism means that repeated queries often result in redundant network requests, increasing latency and consuming API quotas.
+
+Second, the lack of input heterogeneity and robust error handling in batch processing presents a significant bottleneck for high-throughput research. In practical scientific tasks, users often need to process large lists containing mixed identifier types (e.g., names, CIDs, SMILES). Existing tools typically require users to manually classify and preprocess these inputs, adding complexity to the workflow. Moreover, their fault tolerance is limited: a single invalid identifier within a batch of hundreds can often cause the entire task to abort or return incomplete results without explicitly reporting which entries failed. This lack of a structured partial success/failure reporting mechanism makes error tracking difficult and compromises the transparency and reliability of the data acquisition process.
+
[…]
