### 2025-07-17 0a97fda
Update paper.bib
