### 2025-07-17 HzaCode/ChemInformant#2 @HzaCode
Thank you for this excellent suggestion and for providing the reference to the official packaging guide. This is indeed a much better way to handle optional dependencies.

I have implemented this change in the `pyproject.toml` file on the `joss-review-response` branch. I've created a `[plot]` extra that includes `matplotlib` and `Pillow`.

```toml
[project.optional-dependencies]
plot = [
    "matplotlib>=3.5.0",
    "Pillow>=9.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.5"
]
