### 2025-07-17 HzaCode/ChemInformant#4 @HzaCode
Thank you for this essential request.

I have created a new benchmark script (`benchmark.py`) in the repository, which can be run by any user. This script directly addresses your request by comparing three key scenarios:
1.  **`ChemInformant` with a cold cache** (representing the initial batch query).
2.  **`ChemInformant` with a warm cache** (representing a subsequent, repeated query).
3.  **`PubChemPy`'s batch interface** (serving as a baseline for a mainstream, same-platform library).

The results from this benchmark have now been integrated into a new **"Performance Evaluation"** section in the revised JOSS paper. Here is a summary of the findings, as presented in the manuscript:

**Table: Performance benchmark results.**
| Scenario | Time (s) | Speed-up (vs. PubChemPy) |
| :--- | :--- | :--- |
| PubChemPy (batch interface) | 6.50 | 1× (Baseline) |
| **ChemInformant — Cold Cache** | **1.40** | **4.6×** |
| **ChemInformant — Warm Cache** | **0.135** | **48×** |

These results demonstrate that ChemInformant's architecture provides a **4.6× speed-up** over the baseline library even on an initial query, and the built-in caching offers a **48× speed-up** on subsequent queries.

The complete benchmark script can be found on the `joss-review-response` branch.

I believe this provides the concrete data needed to substantiate the performance claims. Thank you for pushing for this higher standard of validation.
