### 2025-07-17 HzaCode/ChemInformant#5 @HzaCode
I have followed your recommendation and completely dropped support for Pydantic v1. The `pyproject.toml` file has been updated to require `pydantic>=2.0`, and all backward-compatibility code has been removed from the library.

Thank you for the guidance on this.
