### 2025-07-17 HzaCode/ChemInformant#6 @HzaCode
Thank you for this insightful feedback on writing style and tone. This is an excellent point, and I agree that making subjective claims about "simplicity" or "ease of use" is not appropriate for a scientific paper. The quality of the software's design should speak for itself.

Following your suggestion, I have carefully reviewed and revised the entire manuscript and all documentation to remove such language. The claims of "Simplicity and Ease of Use" have been completely removed from the `Statement of Need`.

Instead of describing the API as "intuitive," the new version of the paper now focuses on objectively describing its features, such as the layered design and the native Pandas integration. The `Example Usage` section has also been updated to demonstrate this simplicity in practice, allowing the code to make the case on its own.

Thank you for helping me improve the scientific rigor and professionalism of the manuscript.
