### 2025-07-17 HzaCode/ChemInformant#7 @HzaCode
Thank you for this crucial suggestion.

Following your and the other reviewer's advice, I have conducted a broader survey of existing tools across multiple platforms, including other Python clients on PyPI, as well as prominent libraries in the R and Java ecosystems.

The `State of the Field` section in the revised manuscript has been completely rewritten based on this survey. The centerpiece of this new section is a detailed comparative analysis, presented in **Table 1**. This table compares `ChemInformant` against five other libraries (`PubChemPy`, `PubChemR`, `webchem`, `ChemSpiPy`, `PubChem4J`) across eleven key features relevant to modern, automated research workflows, such as:
*   Built-in rate-limiting and retries
*   Native support for mixed-identifier batch queries
*   Batch query fault tolerance
*   Project activity status

This detailed, feature-by-feature comparison now clearly highlights the specific gaps in the existing tool landscape that `ChemInformant` is designed to fill. It also serves as the evidentiary foundation for the arguments made in the revised `Statement of Need`.
