### 2025-07-17 HzaCode/ChemInformant#8 @HzaCode
Thank you for this invaluable and detailed critique.

Inspired by your feedback, I conducted a thorough review of the existing ecosystem, including a survey of other client libraries and an analysis of common user-reported issues and community feedback. Based on this research, I have completely rewritten the `Statement of Need` section.

The new version discards the previous points you critiqued. It now builds its argument on core challenges identified from the community and software review.

The full text of the revised paper is available in the `joss-review-response` branch.

[https://github.com/HzaCode/ChemInformant/blob/joss-review-response/paper/paper.md](https://github.com/HzaCode/ChemInformant/blob/joss-review-response/paper/paper.md)

Thank you again for your guidance, which has fundamentally improved the paper's central argument.
