### 2025-07-17 #8341 @HzaCode
Hi @espottesmith and reviewers,

Thank you for your patience. I am writing to announce that a major revision of `ChemInformant` has been completed, based on the invaluable and comprehensive feedback provided during this review process. All changes have been pushed to a new branch, `joss-review-response`, for your consideration.

The library's public API, internal architecture, and all documentation have been overhauled to address the key concerns that were raised.

Here is a summary of the most significant changes:

### 1. Complete API & Architectural Refactoring (In response to feedback from @cthoyt and @LIVazquezS)

*   **New Layered API Design:** The library now features a dual API: a "convenience layer" of simple, explicit `get_<property>()` functions for quick lookups, and a "power layer" engine (`get_properties`) for high-performance batch operations. This new design directly addresses the feedback regarding API clarity and functional depth.
*   **Native Pandas Integration:** The core `get_properties` function now returns a Pandas DataFrame by default, providing "analysis-ready" data and ensuring seamless integration into the Python data science ecosystem.
*   **Robust Fault Tolerance:** The new batch processing mechanism provides structured status reporting for each identifier, making it easy to track successes and failures without interrupting the entire workflow.

### 2. Strengthened Manuscript & Documentation (In response to feedback from @cthoyt and @LIVazquezS)

*   **Rewritten Statement of Need:** The paper's central argument has been completely rewritten, grounded in a thorough analysis of the limitations of existing tools.
*   **Comprehensive State of the Field:** A detailed comparative analysis (**Table 1**) of six mainstream clients across multiple platforms has been added.
*   **New Performance Benchmark Section:** A new section with a corresponding table (**Table 2**) has been added to the paper, quantitatively demonstrating a **4.6x to 48x performance advantage** over the baseline.
*   **New Interactive User Manual:** A new, comprehensive [Jupyter Notebook User Manual](https://github.com/HzaCode/ChemInformant/blob/joss-review-response/examples/ChemInformant_User_Manual_v2.1.ipynb) has been created to serve as the primary documentation.

### 3. Technical & Packaging Improvements (In response to feedback from @cthoyt)

*   Support for Pydantic v1 has been dropped in favor of `pydantic>=2.0`.
*   Optional dependencies are now managed via `extras` in `pyproject.toml`.

---

To facilitate the review and testing process, the revised version has been published as **`v2.0.0` on PyPI**.

The revised manuscript can be viewed directly on the review branch here:
 [https://github.com/HzaCode/ChemInformant/blob/joss-review-response/paper/paper.md](https://github.com/HzaCode/ChemInformant/blob/joss-review-response/paper/paper.md)

I have also posted individual replies to each of @cthoyt's issues on the repository's issue tracker.

I am looking forward to your thoughts and any further guidance you may have.

Thank you again for the time and guidance you have invested in improving this work.
