### 2025-07-18 #8341 @cthoyt
FYI @HzaCode can you please merge the changes into the main branch? as of now, it's not obvious that these changes got into the repository.

It's also best software practice to have pull requests that address one issue at a time (for future reference). Reviewing the changes that are all lumped together effectively takes as long as the initial review + manual time for figuring out what parts of the super-difff did what, which is unfortunately not very cool for reviewers who are volunteering their time
