### 2025-07-19 76d3143
Update paper.bib
Add citation entry for `requests-cache` v1.2.1 to bibliography
