+++ paper/paper.bib
@@ -129,3 +129,14 @@
   note      = {Accessed 17 Jul 2025}
 }
 
+@software{RequestsCache,
+  author    = {Haritonov, Roman and Cook, John W.},
+  title     = {requests-cache: Persistent HTTP Cache for Python \texttt{requests}},
+  version   = {1.2.1},
+  year      = {2024},
+  month     = jun,
+  publisher = {GitHub},
+  url       = {https://github.com/requests-cache/requests-cache},
+  note      = {Release 1.2.1, accessed 2025-07-19}
+}
+
