### 2025-07-19 f63f206
Update paper.md
polish: refine paper before final merge
