### 2025-07-19 74493fe
Update README.md
finalize cleaned-up README for merge into main
