### 2025-07-19 a067768
Update paper.md
polish: refine paper before final merge
