+++ paper/paper.md
@@ -27,33 +27,33 @@ license: MIT
 
 ## Summary
 
-ChemInformant is a modern, workflow-centric Python client designed for high-throughput and automated PubChem data tasks. Its architecture is engineered to support the direct conversion of various chemical identifiers, including mixed-type lists, into "analysis-ready" Pandas DataFrames [@Pandas], in order to streamline the entire workflow from data retrieval to scientific analysis. As a natively integrated solution, the tool enables zero-configuration robustness by default, incorporating mechanisms such as persistent caching, automatic rate-limiting with exponential backoff, and runtime type validation with Pydantic [@Pydantic]. In benchmark tests, ChemInformant's initial query was 4.6 times faster than the batch interface of a mainstream library; with a warm cache, the performance advantage increased to 48-fold, achieving response times suitable for interactive analysis tools. By systematically addressing key limitations of existing tools in network reliability, batch processing flexibility, and maintainability, ChemInformant is positioned to serve as an efficient and reliable data infrastructure component for the Python cheminformatics ecosystem.
+ChemInformant is a Python client designed for programmatic access to PubChem, with a focus on high-throughput and automated data retrieval tasks. Its architecture facilitates the direct conversion of various chemical identifiers, including mixed-type lists, into analysis-ready Pandas DataFrames [@Pandas], aiming to streamline workflows from data acquisition to analysis. The package integrates several features to enhance operational robustness, such as persistent HTTP caching, automatic rate-limiting with exponential backoff retries, and runtime data validation using Pydantic [@Pydantic]. In benchmark tests comparing batch property retrieval, ChemInformant demonstrated a 4.6-fold performance increase over a widely-used library in initial queries. With caching enabled, this advantage increased to 48-fold, yielding response times suitable for interactive data analysis. By addressing identified limitations in existing tools related to network reliability, batch processing, and maintainability, ChemInformant provides a reliable and efficient component for the Python cheminformatics ecosystem.
 
 ## Statement of Need
 
-Programmatic access to PubChem [@PubChem; @Kim2018PUGREST] has become a critical infrastructure component in modern chemical and life sciences research. As research scales and automated workflows become more prevalent, researchers face three core challenges with existing client tools: the reliability of network access, the flexibility and fault tolerance of batch processing, and the sustainability risks associated with long-term library maintenance.
+Programmatic access to the PubChem database [@PubChem] is a foundational component for many research workflows in chemistry and life sciences. As these workflows become increasingly automated and scaled, researchers encounter recurring challenges with existing client libraries, primarily concerning network reliability, batch processing capabilities, and the long-term sustainability of the tools themselves.
 
-First, network reliability is a primary concern. PubChem's API service enforces strict dynamic rate limits (typically ≤5 requests per second) and returns `HTTP 503 (Server Busy)` errors under high load [@PubChemUsagePolicy]. Mainstream clients like PubChemPy [@PubChemPy] generally lack built-in automatic throttling and exponential backoff retry mechanisms, which makes automated scripts brittle when faced with real-world network fluctuations. Community feedback widely reports such stability issues, with common workarounds still relying on manual insertion of delays. Furthermore, the absence of a persistent caching mechanism means that repeated queries often result in redundant network requests, increasing latency and consuming API quotas.
[…]
