+++ .github/workflows/lint.yml
@@ -1,25 +1,0 @@
-name: Linting
-
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-
-jobs:
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-      - name: Set up Python 3.11
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.11'
-      - name: Install Flake8 (and dependencies if needed by flake8 itself)
-        run: |
-          python -m pip install --upgrade pip
-          pip install ".[dev]"
-      - name: Lint with Flake8
-        run: |
-          flake8 src/ 
