+++ .github/workflows/tests.yml
@@ -9,32 +9,41 @@ on:
 jobs:
   build:
     runs-on: ubuntu-latest
+
     steps:
-    - name: Checkout repository
-      uses: actions/checkout@v4
-    - name: Set up Python 3.11
-      uses: actions/setup-python@v5
-      with:
-        python-version: '3.11'
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install ".[dev,plotting]" pytest-cov
+      - name: Checkout repository
+        uses: actions/checkout@v4
 
-    - name: Verify plotting library installation
-      run: |
-        python -c "import sys; print('Python version:', sys.version); import PIL; print('PIL imported successfully:', PIL.__version__); import matplotlib; print('Matplotlib imported successfully:', matplotlib.__version__); import matplotlib.pyplot; print('matplotlib.pyplot imported successfully')"
+      - name: Set up Python 3.11
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+          cache: "pip"
 
-    - name: Run tests and print coverage report content
-      run: |
-        set -e
-        export PYTHONPATH=$PYTHONPATH:$(pwd)/src
-        echo "PYTHONPATH is now: $PYTHONPATH"
-        pytest --cov=src/ChemInformant --cov-report=xml -v tests/
-        echo "--- Attempting to print contents of coverage.xml ---"
-        cat coverage.xml || echo ">>> coverage.xml file not found! <<<"
-        echo "--- End of coverage.xml content ---"
-    - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v4
-      with:
-        verbose: true 
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install ".[dev,plotting]" pytest-cov black isort
+
+      # ---------- Lint ---------- #
+      - name: Run Black (check)
+        run: black --check src/ tests/
+
+      - name: Run isort (check)
+        run: isort --check src/ tests/
+
+      - name: Run Flake8
+        run: flake8 src/ tests/
+
+      # ---------- Tests ---------- #
+      - name: Run Pytest with coverage
+        env:
+          PYTHONPATH: ${{ github.workspace }}/src
+        run: |
+          pytest --cov=src/ChemInformant --cov-report=xml -v tests/
+
+      # ---------- Codecov ---------- #
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v4
+        with:
+          verbose: true
