+++ .github/workflows/tests.yml
@@ -23,26 +23,24 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install ".[dev,plotting]" pytest-cov black isort
+          pip install ".[dev,plotting]" pytest-cov
 
-      # ---------- Lint ---------- #
-      - name: Run Black (check)
-        run: black --check src/ tests/
+      - name: Verify plotting libraries
+        run: |
+          python - <<'PY'
+          import sys, PIL, matplotlib, matplotlib.pyplot as plt
+          print("Python", sys.version.split()[0])
+          print("Pillow", PIL.__version__)
+          print("matplotlib", matplotlib.__version__)
+          print("pyplot import OK:", plt)
+          PY
 
-      - name: Run isort (check)
-        run: isort --check src/ tests/
-
-      - name: Run Flake8
-        run: flake8 src/ tests/
-
-      # ---------- Tests ---------- #
-      - name: Run Pytest with coverage
+      - name: Run tests with coverage
         env:
           PYTHONPATH: ${{ github.workspace }}/src
         run: |
           pytest --cov=src/ChemInformant --cov-report=xml -v tests/
 
-      # ---------- Codecov ---------- #
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
         with:
