+++ tests/test_api.py
@@ -1,6 +1,11 @@
 from __future__ import annotations
 
-import json, re, sys, types, pytest
+import json
+import re
+import sys
+import types
+
+import pytest
 import ChemInformant as ci
 from ChemInformant import api_helpers as ah
 from ChemInformant import cheminfo_api as capi
@@ -8,14 +13,15 @@ from ChemInformant import models
 
 
 # ------------------------------------------------------------------#
-# 0. Use in-memory HTTP cache
+# 0. Use in-memory HTTP cache for all tests
 @pytest.fixture(autouse=True)
 def _mem_cache():
+    """Ensure all tests use a fresh, in-memory cache for speed and isolation."""
     ci.setup_cache(backend="memory")
 
 
 # ------------------------------------------------------------------#
-# 1. Fake CachedSession (implements cache.disabled())
+# 1. Fake CachedSession (to isolate tests from network)
 class _DummyCtx:
     def __enter__(self): ...
     def __exit__(self, exc_type, exc, tb): ...
@@ -36,7 +42,7 @@ class _Session:
 
 
 # ------------------------------------------------------------------#
-# 2. Fake HTTP response
+# 2. Fake HTTP Response Object
 class _Resp:
     def __init__(self, status=200, data: dict | None = None,
                  from_cache=False,
@@ -53,7 +59,7 @@ class _Resp:
 
 
 # ------------------------------------------------------------------#
-# 3. Stub all network traffic
+# 3. Central Network Stub to Intercept API Calls
 def _fake_execute_fetch(url: str) -> _Resp:                       # noqa: C901
     """Return canned PubChem-like JSON for the URLs used in tests."""
     # name/smiles → CID
@@ -104,15 +110,16 @@ def _fake_execute_fetch(url: str) -> _Resp:                       # noqa: C901
 
 
 # ------------------------------------------------------------------#
-# 4. Wire stubs into api_helpers
+# 4. Wire stubs into the application using a pytest fixture
 @pytest.fixture(autouse=True)
 def _wire_net(monkeypatch):
+    """Fixture to replace all network functions with fakes for all tests."""
     monkeypatch.setattr(ah, "get_session", lambda: _Session())
     monkeypatch.setattr(ah, "_execute_fetch", _fake_execute_fetch)
 
 
 # ------------------------------------------------------------------#
-# 5. Identifier resolution
+# 5. Identifier Resolution Logic Tests
 def test_resolve_variants():
     assert capi._resolve_to_single_cid(2244) == 2244
     assert capi._resolve_to_single_cid("2244") == 2244
@@ -124,7 +131,7 @@ def test_resolve_variants():
 
 
 # ------------------------------------------------------------------#
-# 6. api_helpers round-trip helpers
+# 6. API Helper Function Tests
 def test_helpers_roundtrip():
     assert ah.get_cids_by_name("aspirin") == [2244]
     assert ah.get_cids_by_smiles("C1CC1") == [999]
@@ -135,7 +142,7 @@ def test_helpers_roundtrip():
 
 
 # ------------------------------------------------------------------#
-# 7. High-level public API
+# 7. High-Level Public API Tests
 def test_get_properties_and_compound():
     df = ci.get_properties(["aspirin"],
                            ["canonical_smiles", "molecular_formula", "cas"])
@@ -148,31 +155,58 @@ def test_get_properties_and_compound():
 
 
 # ------------------------------------------------------------------#
-# 8. 503 → retry path (purge cached 503 then retry)
+# 8. Network Retry Logic Test
 def test_retry_path(monkeypatch):
[…]
