+++ .github/workflows/tests.yml
@@ -23,14 +23,18 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install ".[dev,plotting]" pytest-cov
+          pip install ".[dev,plot]" pytest-cov
 
       - name: Verify plotting libraries
         run: |
           python - <<'PY'
-          import sys, PIL, matplotlib, matplotlib.pyplot as plt
+          import sys
+          import matplotlib
+          import matplotlib.pyplot as plt
+          from PIL import __version__ as PILLOW_VERSION
+
           print("Python", sys.version.split()[0])
-          print("Pillow", PIL.__version__)
+          print("Pillow", PILLOW_VERSION)
           print("matplotlib", matplotlib.__version__)
           print("pyplot import OK:", plt)
           PY
