### 2025-07-20 360634e
feat(docs): overhaul documentation with MkDocs and Material
This commit completely overhauls the project documentation by migrating from the previous system to a modern, automated workflow using MkDocs with the Material for MkDocs theme.

Key changes include:
-   Initializes `mkdocs.yml` with a professional configuration.
-   Creates a comprehensive documentation structure with a homepage, usage guide, Jupyter tutorial link, and an automated API reference.
-   Adds detailed docstrings to all public functions and classes in the codebase to enable automatic API documentation generation via `mkdo
[…]
