+++ src/ChemInformant/api_helpers.py
@@ -1,10 +1,16 @@
+"""
+Internal helper module for handling direct communication with the PubChem API.
 
+This module manages the HTTP session, caching, rate-limiting, and retry logic,
+and provides low-level functions that wrap specific PubChem PUG-REST endpoints.
+These functions are not intended for direct use by end-users.
+"""
 from __future__ import annotations
 
 import random
 import sys
 import time
-from typing import List
+from typing import List, Dict, Any
 from urllib.parse import quote
 
 import requests
@@ -27,8 +33,20 @@ def setup_cache(
     cache_name: str = "pubchem_cache",
     backend: str = "sqlite",
     expire_after: int = 7 * 24 * 3600,
-    **kw,
+    **kw: Any,
 ) -> None:
+    """
+    Configures and initializes the persistent cache for API requests.
+
+    This function sets up a `requests_cache` session that will store API
+    responses on disk, significantly speeding up repeated queries.
+
+    Args:
+        cache_name: The name of the cache file (without extension).
+        backend: The cache backend to use (e.g., 'sqlite', 'redis').
+        expire_after: The cache expiration time in seconds. Defaults to one week.
+        **kw: Additional keyword arguments passed to `requests_cache.CachedSession`.
+    """
     global _session
     _session = requests_cache.CachedSession(
         cache_name     = cache_name,
@@ -39,6 +57,7 @@ def setup_cache(
     )
 
 def get_session() -> requests_cache.CachedSession:
+    """Gets the current cached session, initializing it if necessary."""
     global _session
     if _session is None:
         setup_cache()
@@ -46,11 +65,17 @@ def get_session() -> requests_cache.CachedSession:
 
 # Unified fetch with retry and rate-limit
 def _execute_fetch(url: str) -> requests.Response:
+    """Executes a single GET request using the global session."""
     return get_session().get(url, timeout=REQUEST_TIMEOUT)
 
-def _fetch_with_ratelimit_and_retry(url: str):
+def _fetch_with_ratelimit_and_retry(url: str) -> Dict[str, Any] | List[Any] | str | None:
+    """
+    Performs a GET request with rate-limiting and exponential backoff retry logic.
+    This is the core network function that ensures robustness.
+    """
     global last_api_call_time
 
+    # Enforce rate limit
     elapsed = time.time() - last_api_call_time
     if elapsed < MIN_REQUEST_INTERVAL:
         time.sleep(MIN_REQUEST_INTERVAL - elapsed)
@@ -61,91 +86,8 @@ def _fetch_with_ratelimit_and_retry(url: str):
         try:
             resp = _execute_fetch(url)
 
+            # Bust cache for stale 503 errors
             if getattr(resp, "from_cache", False) and resp.status_code == 503:
                 key = get_session().cache.create_key(resp.request)
                 get_session().cache.delete(key)
-                with get_session().cache.disabled():
-                    resp = _execute_fetch(url)
-
-            if resp.status_code in (200, 404):
-                if resp.status_code == 404:
-                    return None
-                ctype = resp.headers.get("Content-Type", "").lower()
-                return resp.json() if "application/json" in ctype else resp.text
-
-            if resp.status_code == 503:
-                print(f"[ChemInformant] 503 → retry in {backoff:.1f}s", file=sys.stderr)
-            else:
-                resp.raise_for_status()
-
-        except requests.exceptions.RequestException as e:
-            print(f"[ChemInformant] network error {e} → retry in {backoff:.1f}s", file=sys.stderr)
-
-        time.sleep(backoff)
[…]
