+++ docs/notebook.md
@@ -1,0 +1,3 @@
+This tutorial demonstrates a practical workflow using ChemInformant within a Jupyter Notebook to analyze a list of common drugs.
+
+You can **[download the full notebook from our GitHub repository](https://github.com/HzaCode/ChemInformant/blob/main/examples/ChemInformant_User_Manual_v1.0.ipynb)** to run it interactively
