+++ .github/workflows/docs.yml
@@ -22,10 +22,17 @@ jobs:
           python-version: 3.11
           cache: "pip"
 
-      - name: Install dependencies
+      - name: Install dependencies with pinned versions
         run: |
           python -m pip install --upgrade pip
-          pip install mkdocs mkdocs-material "mkdocstrings[python]"
+          # THIS IS THE CRITICAL FIX:
+          # We are installing a known-good set of versions to avoid
+          # any dependency conflicts that might break the plugins.
+          pip install \
+            "mkdocs==1.5.3" \
+            "mkdocs-material==9.5.15" \
+            "mkdocstrings==0.24.1" \
+            "mkdocstrings-python==1.8.0"
 
       - name: Deploy to GitHub Pages
         run: python -m mkdocs gh-deploy --force
