+++ docs/index.md
@@ -1,6 +1,8 @@
-# <!-- an empty title -->
+
 
 <img src="assets/logo.png" alt="ChemInformant Logo" align="right" width="180" />
+
+# <!-- an empty title -->
 
 ## A Robust and Workflow-Centric Python Client for High-Throughput PubChem Access
 
