### 2025-07-20 b92c29a
Update api_helpers.py
