+++ CONTRIBUTING.md
@@ -1,28 +1,39 @@
+
 # Contributing to ChemInformant
 
-First off, thank you for considering contributing to ChemInformant! We welcome contributions from everyone. Whether it's reporting a bug, suggesting an enhancement, improving documentation, or writing code, your help is valuable.
+First off, thank you for considering contributing to ChemInformant! We welcome and value contributions from everyone. Whether it's reporting a bug, suggesting an enhancement, seeking support, or writing code, your help is appreciated.
 
-This document provides guidelines for contributing to make the process smooth for everyone involved.
+This document provides guidelines for contributing to make the process as smooth as possible for everyone involved.
 
 ## Table Of Contents
 
 - [Ways to Contribute](#ways-to-contribute)
+- [Seeking Support or Asking Questions](#seeking-support-or-asking-questions)
 - [Reporting Bugs](#reporting-bugs)
 - [Suggesting Enhancements](#suggesting-enhancements)
 - [Your First Code Contribution](#your-first-code-contribution)
   - [Setting Up Your Development Environment](#setting-up-your-development-environment)
   - [Code Style](#code-style)
   - [Running Tests](#running-tests)
-  - [Writing Documentation](#writing-documentation)
   - [Submitting Pull Requests](#submitting-pull-requests)
 - [Code of Conduct](#code-of-conduct)
 
 ## Ways to Contribute
 
+- **Ask for Help:** If you have questions about how to use the project, please open an issue.
 - **Report Bugs:** If you find a bug, please report it using [GitHub Issues](https://github.com/HzaCode/ChemInformant/issues).
 - **Suggest Enhancements:** Have an idea for a new feature or an improvement? Open an issue to discuss it.
-- **Write Documentation:** Good documentation is crucial. Help us improve the docs, tutorials, or examples.
 - **Submit Pull Requests:** Contribute code to fix bugs or add new features.
+
+## Seeking Support or Asking Questions
+
+If you have a question about how to use ChemInformant or are not sure if you've encountered a bug, the best place to ask is the [GitHub Issues page](https://github.com/HzaCode/ChemInformant/issues).
+
+When you open an issue for support, please be sure to:
+1.  **Search existing issues** first to see if your question has already been answered.
+2.  Provide a clear and descriptive title.
+3.  Explain what you are trying to achieve and what problem you are facing.
+4.  If possible, include a small, self-contained code snippet to illustrate your question.
 
 ## Reporting Bugs
 
@@ -34,7 +45,7 @@ If it hasn't, please open a new issue and include the following information:
 - **Python Version:**
 - **Operating System:**
 - **Description of the Bug:** Clearly describe the issue.
-- **Steps to Reproduce:** Provide a minimal code example that triggers the bug.
+- **Steps to Reproduce:** Provide a minimal, complete, and verifiable code example that triggers the bug.
 - **Expected Behavior:** What you expected to happen.
 - **Actual Behavior:** What actually happened (including any error messages or tracebacks).
 
@@ -56,7 +67,7 @@ Ready to contribute code? Here’s how to set up `ChemInformant` for local devel
 1.  **Fork the Repository:** Click the "Fork" button on the [ChemInformant GitHub page](https://github.com/HzaCode/ChemInformant).
 2.  **Clone Your Fork:**
     ```bash
-    git clone [https://github.com/YOUR_USERNAME/ChemInformant.git](https://github.com/YOUR_USERNAME/ChemInformant.git)
+    git clone https://github.com/YOUR_USERNAME/ChemInformant.git
     cd ChemInformant
     ```
 3.  **Create a Virtual Environment:** It's highly recommended to use a virtual environment (e.g., `venv`, `conda`).
@@ -65,7 +76,7 @@ Ready to contribute code? Here’s how to set up `ChemInformant` for local devel
     python -m venv .venv
[…]
