+++ README.md
@@ -120,7 +120,6 @@ This project is licensed under the MIT License - see the [LICENSE](LICENSE) file
 ### Citation
 
 If you use **ChemInformant** in your research, please cite **the software** 
-> **Software**  
-> He, Z. *ChemInformant* (Version&nbsp;2.0.0) [Software]. GitHub repository,  
-> https://github.com/HzaCode/ChemInformant.
+> **Software**: He, Z. *ChemInformant* (v2.0.0) [Software]. GitHub, https://github.com/HzaCode/ChemInformant.  
+
 
