### 2025-07-21 HzaCode/ChemInformant#10 @cthoyt
bug noticable in quickstart
part of second round of review for RE: https://github.com/openjournals/joss-reviews/issues/8341

https://github.com/HzaCode/ChemInformant?tab=readme-ov-file#quick-start

shows the table that has CIDs as floats. These should be explicitly handled as strings, not even as ints.
