### 2025-07-21 96c07fc
Delete docs directory
