+++ docs/api.md
@@ -1,30 +1,0 @@
-# API Reference
-
-This page provides a complete reference for all public classes and functions in the `ChemInformant` library. The content is automatically generated from the docstrings in the source code.
-
-## High-Level API
-
-These are the main functions and classes intended for direct, top-level use.
-
-::: ChemInformant
-    options:
-      show_root_toc_entry: false
-      heading_level: 2
-      members_order: source
-    members:
-      - get_properties
-      - get_compound
-      - get_compounds
-      - get_weight
-      - get_formula
-      - get_cas
-      - get_iupac_name
-      - get_canonical_smiles
-      - get_isomeric_smiles
-      - get_xlogp
-      - get_synonyms
-      - draw_compound
-      - setup_cache
-      - Compound
-      - NotFoundError
-      - AmbiguousIdentifierError
