### 2025-07-21 37029fc
Delete mkdocs.yml
