+++ mkdocs.yml
@@ -1,52 +1,0 @@
-site_name: ChemInformant
-site_description: 'A Robust and Workflow-Centric Python Client for High-Throughput PubChem Access'
-site_author: 'Ang'
-site_url: 'https://hezhiang.com/ChemInformant/'
-
-repo_name: 'HzaCode/ChemInformant'
-repo_url: 'https://github.com/HzaCode/ChemInformant'
-edit_uri: 'edit/main/docs/'
-
-nav:
-  - Home: index.md
-  - Usage Guide: usage.md
-  - Jupyter Tutorial: notebook.md
-  #- API Reference: api.md
-
-theme:
-  name: material
-  palette:
-    - scheme: default
-      toggle:
-        icon: material/brightness-7
-        name: Switch to dark mode
-    - scheme: slate
-      toggle:
-        icon: material/brightness-4
-        name: Switch to light mode
-  features:
-    - navigation.tabs
-    - navigation.sections
-    - navigation.top
-    - search.suggest
-    - search.highlight
-    - content.tabs.link
-    - content.code.copy
-    - content.code.annotate
-
-plugins:
-  - search
-  - mkdocstrings:
-      handlers:
-        python:
-          paths: [src]
-
-markdown_extensions:
-  - pymdownx.highlight:
-      anchor_linenums: true
-  - pymdownx.inlinehilite
-  - pymdownx.snippets
-  - pymdownx.superfences
-  - admonition
-  - toc:
-      permalink: true
