+++ .github/workflows/docs.yml
@@ -1,31 +1,0 @@
-name: Deploy Documentation
-
-on:
-  push:
-    branches:
-      - main
-
-permissions:
-  contents: write
-
-jobs:
-  deploy-docs:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.11' 
-          cache: "pip"
-
-      - name: Install latest dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install mkdocs-material "mkdocstrings[python]"
-
-      - name: Deploy to GitHub Pages
-        run: python -m mkdocs gh-deploy --force
