### 2025-07-21 1d25478
fix(api): Ensure CID column is always a string
Casts the 'cid' column to pandas' 'string' dtype at the end of `get_properties`. This prevents it from being incorrectly displayed as a float (e.g., 2244.0) when failed lookups are present in the batch.
