### 2025-07-21 27a8da7
test(api): Add comprehensive tests for CID dtype and coverage
- Verify the 'cid' column is always a string and never a float.
- Increase test coverage by testing convenience functions and edge cases.
