+++ tests/test_api.py
@@ -6,206 +6,180 @@ import sys
 import types
 
 import pytest
+import pandas as pd
 import ChemInformant as ci
 from ChemInformant import api_helpers as ah
 from ChemInformant import cheminfo_api as capi
 from ChemInformant import models
 
+# ==============================================================================
+# MOCKING INFRASTRUCTURE
+# ==============================================================================
 
-# ------------------------------------------------------------------#
-# 0. Use in-memory HTTP cache for all tests
 @pytest.fixture(autouse=True)
 def _mem_cache():
     """Ensure all tests use a fresh, in-memory cache for speed and isolation."""
     ci.setup_cache(backend="memory")
 
-
-# ------------------------------------------------------------------#
-# 1. Fake CachedSession (to isolate tests from network)
 class _DummyCtx:
     def __enter__(self): ...
     def __exit__(self, exc_type, exc, tb): ...
-
-
 class _Cache:
     def create_key(self, _): return "K"
-    def delete(self, _):     ...
-    def disabled(self):      return _DummyCtx()
-
-
+    def delete(self, _): ...
+    def disabled(self): return _DummyCtx()
 class _Session:
-    def __init__(self):
-        self.cache = _Cache()
-
-    def get(self, url, timeout=None):
-        return _fake_execute_fetch(url)
-
-
-# ------------------------------------------------------------------#
-# 2. Fake HTTP Response Object
+    def __init__(self): self.cache = _Cache()
+    def get(self, url, timeout=None): return _fake_execute_fetch(url)
 class _Resp:
-    def __init__(self, status=200, data: dict | None = None,
-                 from_cache=False,
-                 ctype="application/json"):
-        self.status_code = status
-        self.from_cache = from_cache
-        self.headers = {"Content-Type": ctype}
-        self.request = object()
-        self._data = data or {}
-
+    def __init__(self, status=200, data: dict | None = None, from_cache=False, ctype="application/json"):
+        self.status_code, self.from_cache, self.headers, self.request, self._data = status, from_cache, {"Content-Type": ctype}, object(), data or {}
     def json(self): return self._data
     @property
     def text(self): return json.dumps(self._data)
 
-
-# ------------------------------------------------------------------#
-# 3. Central Network Stub to Intercept API Calls
-def _fake_execute_fetch(url: str) -> _Resp:                       # noqa: C901
-    """Return canned PubChem-like JSON for the URLs used in tests."""
-    # name/smiles → CID
+def _fake_execute_fetch(url: str) -> _Resp:
+    """Enhanced mock API to handle more test cases."""
+    # Identifier resolution
     m = re.search(r"/compound/(name|smiles)/([^/]+)/cids", url)
     if m:
-        token = re.sub(
-            r"%..", lambda m: bytes.fromhex(m.group()[1:]).decode(), m.group(2)
-        )
-        mapping = {"aspirin": [2244], "ambiguous": [1, 2], "C1CC1": [999]}
+        token = re.sub(r"%..", lambda m: bytes.fromhex(m.group()[1:]).decode(), m.group(2))
+        mapping = {"aspirin": [2244], "caffeine": [2519], "ambiguous": [1, 2], "C1CC1": [999], "nonexistent": []}
         return _Resp(data={"IdentifierList": {"CID": mapping.get(token, [])}})
-    # batch properties with pagination
+    # Batch properties (with pagination logic)
[…]
