### 2025-07-21 1896e08
docs: Update README example output
Updates the Quick Start example to show the correct string format for CIDs, reflecting the recent bug fix.
