### 2025-07-21 HzaCode/ChemInformant#10 @HzaCode
Thank you for identifying this issue. The pull request that addresses it, **#11**, is now ready for your review.
