### 2025-07-21 a1c47bd
Update api_helpers.py
