### 2025-07-21 43c9cbc
Update cheminfo_api.py
