### 2025-07-21 e31c36f
Update cheminfo_api.py
