+++ src/ChemInformant/cheminfo_api.py
@@ -3,6 +3,7 @@ This module contains the high-level, user-facing API for ChemInformant.
 These functions are designed for ease of use and to support common
 cheminformatics workflows.
 """
+
 from __future__ import annotations
 
 import re
@@ -16,9 +17,11 @@ from .models import Compound, NotFoundError, AmbiguousIdentifierError
 
 _SMILES_TOKENS = re.compile(r"[=#\[\]\(\)]|\d|Br|Cl|Si", re.I)
 
+
 def _looks_like_smiles(s: str) -> bool:
     """Checks if a string contains characters typical of SMILES."""
     return bool(_SMILES_TOKENS.search(s))
+
 
 def _resolve_to_single_cid(identifier: Union[str, int]) -> int:
     """
@@ -48,15 +51,17 @@ def _resolve_to_single_cid(identifier: Union[str, int]) -> int:
 
     raise NotFoundError(identifier)
 
+
 PROPERTY_ALIASES: Dict[str, List[str]] = {
-    "molecular_weight":   ["MolecularWeight"],
-    "molecular_formula":  ["MolecularFormula"],
-    "canonical_smiles":   ["CanonicalSMILES", "ConnectivitySMILES"],
-    "isomeric_smiles":    ["IsomericSMILES", "SMILES"],
-    "iupac_name":         ["IUPACName"],
-    "xlogp":              ["XLogP"],
+    "molecular_weight": ["MolecularWeight"],
+    "molecular_formula": ["MolecularFormula"],
+    "canonical_smiles": ["CanonicalSMILES", "ConnectivitySMILES"],
+    "isomeric_smiles": ["IsomericSMILES", "SMILES"],
+    "iupac_name": ["IUPACName"],
+    "xlogp": ["XLogP"],
 }
 _SPECIAL_PROPS = {"cas", "synonyms"}
+
 
 def get_properties(
     identifiers: Iterable[Union[str, int]],
@@ -66,13 +71,13 @@ def get_properties(
     Retrieves multiple properties for a list of chemical identifiers.
     """
     identifiers = list(identifiers)
-    properties  = list(properties)
+    properties = list(properties)
 
     if not identifiers or not properties:
         return pd.DataFrame()
 
-    regular     = [p for p in properties if p in PROPERTY_ALIASES]
-    specials    = [p for p in properties if p in _SPECIAL_PROPS]
+    regular = [p for p in properties if p in PROPERTY_ALIASES]
+    specials = [p for p in properties if p in _SPECIAL_PROPS]
     unsupported = [p for p in properties if p not in regular + specials]
     if unsupported:
         raise ValueError(f"Unsupported properties: {unsupported}")
@@ -111,7 +116,11 @@ def get_properties(
             api_row = fetched_regular.get(lookup_cid, {})
             for p in regular:
                 val = next(
-                    (api_row[tag] for tag in PROPERTY_ALIASES[p] if tag in api_row and api_row[tag]),
+                    (
+                        api_row[tag]
+                        for tag in PROPERTY_ALIASES[p]
+                        if tag in api_row and api_row[tag]
+                    ),
                     None,
                 )
                 entry[p] = val
@@ -130,42 +139,52 @@ def get_properties(
 
     return df
 
+
 def _scalar(prop: str, identifier: Union[str, int]):
     """Internal helper to get a single value from get_properties."""
     df = get_properties([identifier], [prop])
     return df[prop].iat[0] if not df.empty and df["status"].iat[0] == "OK" else None
 
+
 def get_weight(id_: Union[str, int]) -> float | None:
     """Gets the molecular weight for a single identifier."""
-    return _scalar("molecular_weight",   id_)
+    return _scalar("molecular_weight", id_)
+
 
 def get_formula(id_: Union[str, int]) -> str | None:
[…]
