+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ChemInformant"
-version = "2.0.0"
+version = "2.1.0"
 authors = [
   { name = "Zhiang He", email = "ang@hezhiang.com" },
 ]
