### 2025-07-21 d37bc40
Update paper.md
