### 2025-07-21 d018e34
Update CITATION.cff
