+++ CITATION.cff
@@ -1,4 +1,4 @@
-cff-version: 2.0.0
+cff-version: 2.1.0
 message: "If you use this software, please cite it as below."
 authors:
   - family-names: "He"
