### 2025-07-21 bc07d95
Update CITATION.cff
