### 2025-07-21 1978743
Create LICENSE.md
