### 2025-07-22 47e036b
Update paper.md
