+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ChemInformant"
-version = "2.1.0"
+version = "2.1.1"
 authors = [
   { name = "Zhiang He", email = "ang@hezhiang.com" },
 ]
