### 2025-07-22 a91ed89
Update CITATION.cff
