+++ CITATION.cff
@@ -1,11 +1,11 @@
-cff-version: 2.1.0
+cff-version: 2.1.1
 message: "If you use this software, please cite it as below."
 authors:
   - family-names: "He"
     given-names: "Zhiang"
     orcid: "https://orcid.org/0009-0009-0171-4578"
 title: "ChemInformant"
-version: 2.1.0
+version: 2.1.1
 date-released: 2025
 url: "https://github.com/HzaCode/ChemInformant"
 license: MIT
