### 2025-07-22 #8341 @HzaCode
Thanks, @espottesmith, for the follow-up!

I just wanted to post a quick update. The valuable feedback so far has inspired a few further enhancements:

- Fixed a bug kindly identified by @cthoyt
- Added a new CLI for easier access.
- Expanded the documentation with more real-world examples (e.g., interaction with analysis tools).
- Increased test coverage to 87%.

This process has been a great learning experience, and I'm ready to address any further feedback when you have a chance. Thanks again, everyone!
