+++ src/ChemInformant/__init__.py
@@ -6,49 +6,58 @@ This package provides high-level functions to retrieve chemical data,
 seamlessly handling identifier conversion, batch processing, and network
 robustness through built-in caching and retry mechanisms.
 """
-
 from __future__ import annotations
 
 # Re-export high-level API
-from .cheminfo_api import (  # noqa: F401
-    get_properties,
-    get_compound,
-    get_compounds,
-    get_weight,
-    get_formula,
-    get_cas,
-    get_iupac_name,
-    get_canonical_smiles,
-    get_isomeric_smiles,
-    get_xlogp,
-    get_synonyms,
-    draw_compound,
+from .cheminfo_api import (
+get_properties,
+get_compound,
+get_compounds,
+get_weight,
+get_formula,
+get_cas,
+get_iupac_name,
+get_canonical_smiles,
+get_isomeric_smiles,
+get_xlogp,
+get_synonyms,
+draw_compound,
 )
 
+# --- NEW: Import the new SQL function ---
+from .sql import df_to_sql
+
 # Cache setup
-from .api_helpers import setup_cache  # noqa: F401
+from .api_helpers import setup_cache
 
 # Models and exceptions
-from .models import Compound, NotFoundError, AmbiguousIdentifierError  # noqa: F401
+from .models import Compound, NotFoundError, AmbiguousIdentifierError
 
-from . import api_helpers  # noqa: F401
+from . import api_helpers
 
+# --- NEW: Add df_to_sql to the public API list ---
 __all__ = [
-    "get_properties",
-    "get_compound",
-    "get_compounds",
-    "get_weight",
-    "get_formula",
-    "get_cas",
-    "get_iupac_name",
-    "get_canonical_smiles",
-    "get_isomeric_smiles",
-    "get_xlogp",
-    "get_synonyms",
-    "draw_compound",
-    "setup_cache",
-    "api_helpers",
-    "Compound",
-    "NotFoundError",
-    "AmbiguousIdentifierError",
+# Core functions
+"get_properties",
+"get_compound",
+"get_compounds",
+# Scalar getters
+"get_weight",
+"get_formula",
+"get_cas",
+"get_iupac_name",
+"get_canonical_smiles",
+"get_isomeric_smiles",
+"get_xlogp",
+"get_synonyms",
+# New persistence function
+"df_to_sql",
+# Utilities
+"draw_compound",
+"setup_cache",
+# Objects and Exceptions
+"api_helpers",
+"Compound",
+"NotFoundError",
+"AmbiguousIdentifierError",
 ]
