### 2025-07-24 2e332aa
Update cli.py
