+++ src/ChemInformant/__init__.py
@@ -6,22 +6,23 @@ This package provides high-level functions to retrieve chemical data,
 seamlessly handling identifier conversion, batch processing, and network
 robustness through built-in caching and retry mechanisms.
 """
+
 from __future__ import annotations
 
 # Re-export high-level API
 from .cheminfo_api import (
-get_properties,
-get_compound,
-get_compounds,
-get_weight,
-get_formula,
-get_cas,
-get_iupac_name,
-get_canonical_smiles,
-get_isomeric_smiles,
-get_xlogp,
-get_synonyms,
-draw_compound,
+    get_properties,
+    get_compound,
+    get_compounds,
+    get_weight,
+    get_formula,
+    get_cas,
+    get_iupac_name,
+    get_canonical_smiles,
+    get_isomeric_smiles,
+    get_xlogp,
+    get_synonyms,
+    draw_compound,
 )
 
 # --- NEW: Import the new SQL function ---
@@ -37,27 +38,27 @@ from . import api_helpers
 
 # --- NEW: Add df_to_sql to the public API list ---
 __all__ = [
-# Core functions
-"get_properties",
-"get_compound",
-"get_compounds",
-# Scalar getters
-"get_weight",
-"get_formula",
-"get_cas",
-"get_iupac_name",
-"get_canonical_smiles",
-"get_isomeric_smiles",
-"get_xlogp",
-"get_synonyms",
-# New persistence function
-"df_to_sql",
-# Utilities
-"draw_compound",
-"setup_cache",
-# Objects and Exceptions
-"api_helpers",
-"Compound",
-"NotFoundError",
-"AmbiguousIdentifierError",
+    # Core functions
+    "get_properties",
+    "get_compound",
+    "get_compounds",
+    # Scalar getters
+    "get_weight",
+    "get_formula",
+    "get_cas",
+    "get_iupac_name",
+    "get_canonical_smiles",
+    "get_isomeric_smiles",
+    "get_xlogp",
+    "get_synonyms",
+    # New persistence function
+    "df_to_sql",
+    # Utilities
+    "draw_compound",
+    "setup_cache",
+    # Objects and Exceptions
+    "api_helpers",
+    "Compound",
+    "NotFoundError",
+    "AmbiguousIdentifierError",
 ]
