### 2025-07-24 b019cf4
Update test_sql.py
