### 2025-07-24 ad804cd
Update test_sql.py
