### 2025-07-24 HzaCode/ChemInformant#17 @cthoyt
Documentation is missing full-fledged tutorial on caching
part of second round of review for RE: https://github.com/openjournals/joss-reviews/issues/8341
