### 2025-07-24 HzaCode/ChemInformant#18 @cthoyt
Tutorial link is wrong
part of second round of review for RE: https://github.com/openjournals/joss-reviews/issues/8341

on https://hezhiang.com/ChemInformant/tutorials.html, the link in there goes to a raw page, not the nice formatted jupyter notebook viewer like it should
