### 2025-07-24 HzaCode/ChemInformant#19 @cthoyt
Table in the paper is not readable
Second round of review at https://github.com/openjournals/joss-reviews/issues/8341

As a reader, I'm totally overwhelmed by the table on page 2 and 3 of https://raw.githubusercontent.com/openjournals/joss-papers/joss.08341/joss.08341/10.21105.joss.08341.pdf.

There are a few ways you could go about fixing this:

1. Swap out the multi-line cells for key features for numbers, and use an ordered list in the table caption or in the prose to move this text out
2. don't add citations in software here, this should be somewhere else
3. consider rotating 90 degrees and putting on at the end of the manuscript?
