### 2025-07-24 HzaCode/ChemInformant#20 @cthoyt
Tone down promise of maintainability
Second round of review at https://github.com/openjournals/joss-reviews/issues/8341

> Finally, the maintenance status of client tools poses a potential risk to research reproducibility43
and stability. For instance, PubChemPy, a widely used library in the Python ecosystem, has44
not had a new release since version 1.0.4 in 2017, leaving community-requested features45
unimplemented. More importantly, a lack of active maintenance means that a tool cannot46
adapt to the evolution of the PubChem API itself. Users are often forced to supplement these47
libraries with other tools or custom logic, but such non-integrated solutions typically fail to48
systemically address the aforementioned stability and efficiency issues.

it might be the case that pubchempy isn't maintained anymore, but making a promise that a new package will be maintained is mostly hollow. Your project is just run by you, and it hasn't been for a very long time so there isn't really a reason to believe that you will maintain it for a long time.

therefore, you should tone down this kind of language. focus on how your package could be a drop-in replacement in many places.

One thing you could think about is making a governance model for onboarding new maintainers, a succession plan for moving the code to an organization if you ever get bored with maintaining it, or inviting hard forks in the case. But I think this is getting away from the point, any code published in an academic journal is probably susceptible to being abandoned.
