+++ src/ChemInformant/api_helpers.py
@@ -1,98 +1,114 @@
-"""
-ChemInformant – low‑level helper functions for PubChem PUG‑REST.
+"""Internal helper module for handling direct communication with the PubChem API.
 
-• Manages HTTP session, persistent caching, rate‑limit, retries.  
-• Since v2.3.0 the cache path is resolved via PyStow:
-  ~/.data/cheminformant/cache/pubchem_cache.sqlite (overridable via env vars).
+This module manages the HTTP session, caching, rate-limiting, and retry logic,
+and provides low-level functions that wrap specific PubChem PUG-REST endpoints.
+These functions are not intended for direct use by end-users, but are consumed
+by the main API interface.
 """
-
 from __future__ import annotations
 
-import secrets
+import random
 import sys
 import time
-from datetime import timedelta
-from pathlib import Path
-from typing import Any, Dict, List
+from typing import List, Dict, Any
+from urllib.parse import quote
 
 import requests
 import requests_cache
 
-# ──────────────────────────────────────────────────────────────────────────────
-#  Cache path resolution (PyStow-aware; fallback supported)
-# ──────────────────────────────────────────────────────────────────────────────
-try:
-    import pystow
-except ImportError:
-    pystow = None
-
-
-def _resolve_cache_path(cache_name: str) -> str | Path:
-    if pystow is not None:
-        module = pystow.module("cheminformant")
-        return module.join("cache", name=f"{cache_name}.sqlite")
-    return cache_name
-
-
-# ──────────────────────────────────────────────────────────────────────────────
-#  Constants
-# ──────────────────────────────────────────────────────────────────────────────
+# --- Module Constants ---
 PUBCHEM_API_BASE = "https://pubchem.ncbi.nlm.nih.gov/rest/pug"
-PUG_VIEW_BASE = "https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data"
-REQUEST_TIMEOUT = 15
+PUG_VIEW_BASE    = "https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data"
+REQUEST_TIMEOUT  = 15
 
-MAX_RETRIES = 5
-INITIAL_BACKOFF = 1
-MAX_BACKOFF = 16
-REQUEST_RATE_LIMIT = 5
+MAX_RETRIES, INITIAL_BACKOFF, MAX_BACKOFF = 5, 1, 16
+REQUEST_RATE_LIMIT = 5  # Requests per second
 MIN_REQUEST_INTERVAL = 1.0 / REQUEST_RATE_LIMIT
 
+# --- Module Globals ---
 last_api_call_time: float = 0.0
 _session: requests_cache.CachedSession | None = None
 
 
-# ──────────────────────────────────────────────────────────────────────────────
-#  Session & caching
-# ──────────────────────────────────────────────────────────────────────────────
+# --- Session & Caching ---
+
 def setup_cache(
     cache_name: str = "pubchem_cache",
     backend: str = "sqlite",
-    expire_after: int | float | str | timedelta = 7 * 24 * 3600,
+    expire_after: int = 7 * 24 * 3600,
     **kw: Any,
 ) -> None:
-    """Configure the global CachedSession."""
+    """
+    Configures and initializes the persistent cache for API requests.
+
+    This function sets up a ``requests_cache`` session that will store API
+    responses on disk, significantly speeding up repeated queries and reducing
+    network traffic.
+
+    :param cache_name: The name of the cache file (without extension).
+    :type cache_name: str
+    :param backend: The cache backend to use (e.g., 'sqlite', 'redis').
+    :type backend: str
+    :param expire_after: Cache expiration time in seconds. Defaults to one week.
+    :type expire_after: int
[…]
