+++ docs/source/api_reference.rst
@@ -1,0 +1,12 @@
+API 参考
+=======
+
+本节展示 ChemInformant 提供的所有可编程 API，包括底层查询模块、数据模型等。
+
+.. toctree::
+   :maxdepth: 2
+
+   api_guide
+   api/api_helpers
+   api/cheminfo_api
+   api/models
