+++ docs/source/tutorials.rst
@@ -1,0 +1,12 @@
+=========
+Tutorials
+=========
+
+
+
+
+For the best learning experience, we highly recommend downloading the accompanying Jupyter Notebook for this page. It contains all the runnable code examples, allowing you to get hands-on, modify parameters, and observe the results to gain a deeper understanding of how to apply ChemInformant in practical workflows.
+
+**Download the interactive tutorial:**
+
+`ChemInformant User Manual (ChemInformant_User_Manual_v1.0.ipynb) <https://raw.githubusercontent.com/HzaCode/ChemInformant/main/examples/ChemInformant_User_Manual_v1.0.ipynb>`__

[…]
