+++ docs/requirements.txt
@@ -1,0 +1,5 @@
+sphinx
+sphinx-rtd-theme
+
+# myst-parser
+# sphinx-copybutton
