+++ .github/workflows/docs-deploy.yml
@@ -30,15 +30,14 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install sphinx sphinx-rtd-theme
-          # If you use a requirements file, replace the line above with:
-          # python -m pip install -r docs/requirements.txt
+          # Ensure you have a requirements.txt in your docs folder
+          python -m pip install -r docs/requirements.txt
 
       - name: Build Sphinx documentation
-        run: sphinx-build -b html source _build/html
+        run: sphinx-build -b html docs/source docs/_build/html
 
       - name: Create .nojekyll file
-        run: touch _build/html/.nojekyll
+        run: touch docs/_build/html/.nojekyll
 
       - name: Setup Pages
         uses: actions/configure-pages@v5
@@ -46,7 +45,7 @@ jobs:
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
         with:
-          path: './_build/html'
+          path: './docs/_build/html'
 
       - name: Deploy to GitHub Pages
         id: deployment
