+++ .github/workflows/docs-deploy.yml
@@ -32,6 +32,7 @@ jobs:
           python -m pip install --upgrade pip
           # Ensure you have a requirements.txt in your docs folder
           python -m pip install -r docs/requirements.txt
+          python -m pip install .
 
       - name: Build Sphinx documentation
         run: sphinx-build -b html docs/source docs/_build/html
