### 2025-07-25 HzaCode/ChemInformant#17 @HzaCode
Thanks for your feedback. I've now added a detailed tutorial on caching.

https://hezhiang.com/ChemInformant/caching_guide.html
