### 2025-07-25 fa5f9e1
Update paper.md
