### 2025-07-25 HzaCode/ChemInformant#19 @HzaCode
Thank you for the detailed and very constructive feedback.

I have now implemented a combination of your first two recommendations, which I believe effectively resolves the readability issue.

1.  **Following your first suggestion**, I have streamlined Table 1 significantly. The multi-line descriptions have been removed and replaced with superscripts (¹, ², ³) that correspond to a new, compact notes section below the table.
2.  **In line with your second point**, I have also removed extraneous details like version numbers from the table. The key information about project inactivity dates has been moved into the main prose just before the table, where it fits more naturally.
