+++ paper/paper.md
@@ -27,19 +27,19 @@ license: MIT
 
 ## Summary
 
-ChemInformant is a Python client designed for programmatic access to PubChem, with a focus on high-throughput and automated data retrieval tasks. Its architecture facilitates the direct conversion of various chemical identifiers, including mixed-type lists, into analysis-ready Pandas DataFrames [@Pandas], aiming to streamline workflows from data acquisition to analysis. The package integrates several features to enhance operational robustness, such as persistent HTTP caching, automatic rate-limiting with exponential backoff retries, and runtime data validation using Pydantic [@Pydantic]. In benchmark tests comparing batch property retrieval, ChemInformant demonstrated a 4.6-fold performance increase over a widely-used library in initial queries. With caching enabled, this advantage increased to 48-fold, yielding response times suitable for interactive data analysis. By addressing identified limitations in existing tools related to network reliability, batch processing, and maintainability, ChemInformant provides a reliable and efficient component for the Python cheminformatics ecosystem.
+ChemInformant is a Python client engineered for programmatic access to PubChem, specifically targeting high-throughput and automated data retrieval tasks. Its architecture streamlines the entire workflow from data acquisition to analysis by directly converting large, mixed-type lists of chemical identifiers into analysis-ready Pandas DataFrames [@Pandas]. To ensure operational resilience, the package natively integrates a suite of robustness features, including persistent HTTP caching, automatic rate-limiting with exponential backoff retries, and runtime data validation using Pydantic [@Pydantic]. By systematically addressing critical limitations in existing tools—such as network instability and inefficient batch processing—and offering up to a 48-fold performance increase, ChemInformant delivers a significantly more reliable and efficient component for the modern Python cheminformatics ecosystem.
 
 ## Statement of Need
 
-Programmatic access to the PubChem database [@PubChem] is a foundational component for many research workflows in chemistry and life sciences. As these workflows become increasingly automated and scaled, researchers encounter recurring challenges with existing client libraries, primarily concerning network reliability, batch processing capabilities, and the long-term sustainability of the tools themselves.
+As these workflows become increasingly automated and scaled, researchers encounter recurring challenges with existing client libraries, primarily concerning network reliability, batch processing capabilities, and a lack of workflow-centric API design.
 
-First, network stability is a significant operational concern. The PubChem API service [@Kim2018PUGREST] enforces dynamic rate limits (e.g., ≤5 requests per second) and may return `HTTP 503 (Server Busy)` errors during periods of high traffic [@PubChemUsagePolicy]. Many existing clients, such as PubChemPy [@PubChemPy], do not include built-in mechanisms for automatic request throttling or exponential backoff retries. This can lead to script fragility in automated environments, often requiring users to implement manual delays. Furthermore, the general absence of a persistent caching layer results in redundant network requests for repeated queries, which increases latency and unnecessarily consumes API usage quotas.
+First, network stability is a significant operational concern. The PubChem API service [@Kim2018PUGREST] enforces dynamic rate limits (e.g., ≤5 requests per second) and may return HTTP 503 (Server Busy) errors during periods of high traffic [@PubChemUsagePolicy]. Many existing clients, such as PubChemPy [@PubChemPy], do not include built-in mechanisms for automatic request throttling or exponential backoff retries. This can lead to script fragility in automated environments, often requiring users to implement manual delays. Furthermore, the general absence of a persistent caching layer results in redundant network requests for repeated queries, which increases latency and unnecessarily consumes API usage quotas.
 
[…]
