### 2025-07-25 #8341 @HzaCode
Hi @espottesmith and reviewers,

Thank you all again for the detailed feedback. I have completed the revisions based on your suggestions.

Here is a summary of the key updates:

1.  **Added SQL Output:** I've implemented a new feature to support outputting data directly to an SQL database, which is also available via the CLI.
2.  **Expanded Caching Support:** To provide more flexible backend options, I have integrated support for the `pstow` library.
3.  **Enhanced Paper and Documentation:** As recommended, the paper and documentation have been significantly updated to reflect the latest changes.
4.  **Increased Test Coverage:** The project's test coverage has now been increased to **90%**.

Please take a look whenever you have a moment:)

Thanks again for your invaluable time and guidance through this process.
