### 2025-07-25 ac2d82e
Update cli.rst
