+++ docs/source/cli.rst
@@ -7,12 +7,10 @@ ChemInformant provides a suite of command-line interface (CLI) tools, designed t
 .. contents::
    :local:
 
-
-
 `chemfetch`
 -----------
 
-`chemfetch` is ChemInformant's core data retrieval tool. Its primary responsibility is to accept one or more chemical identifiers provided by the user, initiate a request to the PubChem API to fetch specified chemical properties, and return the results to standard output in a user-selected format. The tool can automatically recognize various identifier types and robustly handle batch requests that include invalid identifiers.
+`chemfetch` is ChemInformant's core data retrieval tool. Its primary responsibility is to accept one or more chemical identifiers provided by the user, initiate a request to the PubChem API to fetch specified chemical properties, and return the results to standard output or a file in a user-selected format. The tool can automatically recognize various identifier types and robustly handle batch requests that include invalid identifiers.
 
 **Usage**
 
@@ -26,9 +24,9 @@ ChemInformant provides a suite of command-line interface (CLI) tools, designed t
 
    One or more required chemical identifiers, separated by spaces. The internal logic of `chemfetch` will attempt to parse the type of each identifier:
 
-   *   **Name:** e.g., ``aspirin``, ``caffeine``, ``water``. The tool uses these names to search in PubChem.
-   *   **PubChem Compound ID (CID):** e.g., ``2244`` for aspirin. This is a specific and unambiguous identifier.
-   *   **SMILES String:** e.g., ``"CC(=O)Oc1ccccc1C(=O)O"``. This is a linear notation for describing a compound's structure.
+   * **Name:** e.g., ``aspirin``, ``caffeine``, ``water``. The tool uses these names to search in PubChem.
+   * **PubChem Compound ID (CID):** e.g., ``2244`` for aspirin. This is a specific and unambiguous identifier.
+   * **SMILES String:** e.g., ``"CC(=O)Oc1ccccc1C(=O)O"``. This is a linear notation for describing a compound's structure.
 
    .. note::
       When a provided SMILES string contains characters that might be interpreted by the shell as special operators (e.g., `(`, `)`, `=`, `#`), it is strongly recommended to enclose the entire string in single or double quotes to ensure it is passed as a single, complete argument to `chemfetch`.
@@ -37,178 +35,131 @@ ChemInformant provides a suite of command-line interface (CLI) tools, designed t
 
    A comma-separated list of properties to precisely specify which data to retrieve for each identifier. If the user does not provide this option, `chemfetch` will use a default set of properties: ``cas,molecular_weight,iupac_name``.
 
-   **The complete list of available properties includes:**
+   The complete list of available properties includes:
 
-   *   ``cas``: CAS Registry Number, a common chemical identifier.
-   *   ``molecular_weight``: Molecular weight, in g/mol.
-   *   ``molecular_formula``: Molecular formula, indicating the number of atoms of each element in the compound.
-   *   ``canonical_smiles``: Canonical SMILES string, a standardized, unique SMILES representation.
-   *   ``isomeric_smiles``: Isomeric SMILES string, which can represent specific stereoisomers.
-   *   ``iupac_name``: The systematic name established by the International Union of Pure and Applied Chemistry (IUPAC).
-   *   ``xlogp``: The theoretically calculated octanol-water partition coefficient, often used to assess a drug's lipophilicity.
-   *   ``synonyms``: A list containing all known synonyms for the compound, such as trade names, common names, etc.
[…]
