### 2025-07-26 #8341 @HzaCode
@cthoyt Thank you for your patience. Your detailed review has been incredibly helpful, and I will start working on the PDF issue now.
