### 2025-07-26 d3d6825
Update paper.md
