### 2025-07-26 63a0d22
Update paper.bib
