+++ paper/paper.bib
@@ -66,17 +66,6 @@
   pmid    = {29718389},
   pmcid   = {PMC6031027}
 }
-
-@software{RequestsCache,
-  author    = {{Requests-Cache Developers}},
-  title     = {requests-cache: Transparent Persistent Cache for Python Requests},
-  version   = {1.2.1},
-  year      = {2024},
-  month     = jun,
-  url       = {https://github.com/requests-cache/requests-cache/releases/tag/v1.2.1},
-  note      = {GitHub release v1.2.1, accessed 17 Jul 2025}
-}
-
 @software{Pydantic,
   author    = {Colvin, Samuel and others},
   title     = {Pydantic},
