### 2025-07-26 9e63ce1
Update paper.md
