### 2025-07-26 bb9f60c
Update paper.bib
