+++ paper/paper.bib
@@ -138,3 +138,10 @@
   publisher = {Zenodo},
   doi       = {10.5281/zenodo.15720682},
   url       = {https://doi.org/10.5281/zenodo.15720682}
+}
+
+
+
+
+
+
