### 2025-07-26 64631bd
Update paper.md
