### 2025-07-26 21c30f8
Update paper.md
