+++ .github/workflows/tests.yml
@@ -9,7 +9,6 @@ on:
 jobs:
   build:
     runs-on: ubuntu-latest
-
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
@@ -45,8 +44,11 @@ jobs:
         run: |
           pytest --cov=src/ChemInformant --cov-report=xml -v tests/
 
+     
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
         with:
-          token: ${{ secrets.CODECOV_TOKEN }}
-          # slug: HzaCode/ChemInformant (this is optional)
+          token: ${{ secrets.CODECOV_TOKEN }} 
+          file: ./coverage.xml              
+          fail_ci_if_error: true            
+     
