+++ pyproject.toml
@@ -77,3 +77,6 @@ dev = [
 
 [tool.setuptools.packages.find]
 where = ["src"]
+
+[tool.coverage.run]
+source = ["ChemInformant"]
