+++ codecov.yml
@@ -1,16 +1,18 @@
 coverage:
-  range: "60...79.99"
+  range: "60..79.99"
   status:
     project:
       default:
         target: 80%
         threshold: 1%
+        informational: false
     patch:
       default:
         target: 85%
+        informational: false
 
 fixes:
-  - "^([^/]+)\\.py$::src/ChemInformant/\\1.py"
+  - "src/::"
 
 comment:
   layout: "reach,diff,flags,files,footer"
