+++ .github/workflows/tests.yml
@@ -2,9 +2,9 @@ name: Tests
 
 on:
   push:
-    branches: [ main ]
+    branches: [main]
   pull_request:
-    branches: [ main ]
+    branches: [main]
 
 jobs:
   build:
@@ -17,25 +17,21 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: "3.11"
-          cache: "pip"
+          cache: pip
 
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install ".[dev,plot]" pytest-cov
+          pip install ".[dev,plot]" pytest-cov coverage-badge
 
       - name: Verify plotting libraries
         run: |
           python - <<'PY'
-          import sys
-          import matplotlib
-          import matplotlib.pyplot as plt
-          from PIL import __version__ as PILLOW_VERSION
-
+          import sys, matplotlib
+          from PIL import __version__ as PILVER
           print("Python", sys.version.split()[0])
-          print("Pillow", PILLOW_VERSION)
+          print("Pillow", PILVER)
           print("matplotlib", matplotlib.__version__)
-          print("pyplot import OK:", plt)
           PY
 
       - name: Run tests with coverage
@@ -44,11 +40,15 @@ jobs:
         run: |
           pytest --cov=src/ChemInformant --cov-report=xml -v tests/
 
-     
-      - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v4
+      - name: Generate coverage badge
+        run: coverage-badge -o coverage.svg -f
+
+      - name: Deploy badge to gh-pages
+        uses: peaceiris/actions-gh-pages@v4
         with:
-          token: ${{ secrets.CODECOV_TOKEN }} 
-          file: ./coverage.xml              
-          fail_ci_if_error: true            
-     
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: .
+          publish_branch: gh-pages
+          keep_files: true
+          allow_empty_commit: false
+          commit_message: "Update coverage badge [skip ci]"
