+++ .github/workflows/tests.yml
@@ -6,15 +6,18 @@ on:
   pull_request:
     branches: [main]
 
+permissions:
+  contents: write
+
 jobs:
   build:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
+      - uses: actions/checkout@v4
+        with:
+          persist-credentials: false
 
-      - name: Set up Python 3.11
-        uses: actions/setup-python@v5
+      - uses: actions/setup-python@v5
         with:
           python-version: "3.11"
           cache: pip
@@ -37,8 +40,7 @@ jobs:
       - name: Run tests with coverage
         env:
           PYTHONPATH: ${{ github.workspace }}/src
-        run: |
-          pytest --cov=src/ChemInformant --cov-report=xml -v tests/
+        run: pytest --cov=src/ChemInformant --cov-report=xml -v tests/
 
       - name: Generate coverage badge
         run: coverage-badge -o coverage.svg -f
