### 2025-07-26 5cb5faf
Update coverage badge [skip ci] b3608e143e19855498763652a42f1ff48d893607
