+++ paper/paper.bib
@@ -1,0 +1,150 @@
+@article{PubChem4J,
+  author  = {Southern, Mark R. and Griffin, Patrick R.},
+  title   = {A Java API for working with PubChem datasets},
+  journal = {Bioinformatics},
+  year    = {2011},
+  month   = mar,
+  volume  = {27},
+  number  = {5},
+  pages   = {741--742},
+  doi     = {10.1093/bioinformatics/btq715},
+  url     = {https://doi.org/10.1093/bioinformatics/btq715},
+  pmid    = {21216779},
+  pmcid   = {PMC3105478}
+}
+
+@software{Pandas,
+  author    = {{The pandas development team}},
+  title     = {pandas-dev/pandas: Pandas},
+  version   = {2.3.1},
+  year      = {2025},
+  month     = jul,
+  publisher = {Zenodo},
+  doi       = {10.5281/zenodo.15831829},
+  url       = {https://doi.org/10.5281/zenodo.15831829}
+}
+
+@article{webchem,
+  author  = {Sz{\"o}cs, Eduard and Stirling, Tam{\'a}s and Scott, Eleni R. and Scharm{\"u}ller, Andreas and Sch{\"a}fer, Ralf B.},
+  title   = {webchem: An R Package to Retrieve Chemical Information from the Web},
+  journal = {Journal of Statistical Software},
+  year    = {2020},
+  month   = may,
+  volume  = {93},
+  number  = {13},
+  pages   = {1--17},
+  doi     = {10.18637/jss.v093.i13},
+  url     = {https://doi.org/10.18637/jss.v093.i13}
+}
+
+@article{PubChem,
+  author  = {Kim, Sunghwan and Chen, Jie and Cheng, Tiejun and Gindulyte, Asta and He, Jia and He, Siqian and Li, Qingliang and Shoemaker, Benjamin A. and Thiessen, Paul A. and Yu, Bo and Zaslavsky, Leonid and Zhang, Jian and Bolton, Evan E.},
+  title   = {PubChem 2023 Update},
+  journal = {Nucleic Acids Research},
+  year    = {2023},
+  month   = jan,
+  volume  = {51},
+  number  = {D1},
+  pages   = {D1373--D1380},
+  doi     = {10.1093/nar/gkac956},
+  url     = {https://doi.org/10.1093/nar/gkac956},
+  pmid    = {36305812},
+  pmcid   = {PMC9825602}
+}
+
+@article{Kim2018PUGREST,
+  author  = {Kim, Sunghwan and Thiessen, Paul A. and Cheng, Tiejun and Yu, Bo and Bolton, Evan E.},
+  title   = {An Update on PUG-REST: RESTful Interface for Programmatic Access to PubChem},
+  journal = {Nucleic Acids Research},
+  year    = {2018},
+  month   = jul,
+  volume  = {46},
+  number  = {W1},
+  pages   = {W563--W570},
+  doi     = {10.1093/nar/gky294},
+  url     = {https://doi.org/10.1093/nar/gky294},
+  pmid    = {29718389},
+  pmcid   = {PMC6031027}
+}
+@software{Pydantic,
+  author    = {Colvin, Samuel and others},
+  title     = {Pydantic},
+  version   = {2.11.7},
+  year      = {2025},
+  month     = jun,
+  publisher = {Zenodo},
+  doi       = {10.5281/zenodo.15662245},
+  url       = {https://doi.org/10.5281/zenodo.15662245}
+}
+
+@software{PubChemPy,
+  author    = {Swain, Matt},
+  title     = {PubChemPy v1.0.4},
+  version   = {1.0.4},
+  year      = {2017},
+  month     = apr,
+  publisher = {Zenodo},
[…]
