+++ paper/paper.md
@@ -1,0 +1,114 @@
+---
+title: 'ChemInformant: A Robust and Workflow-Centric Python Client for High-Throughput PubChem Access'
+tags:
+  - python
+  - pubchem
+  - chemistry
+  - api
+  - pandas
+  - cache
+  - pydantic
+  - cheminformatics
+  - data science
+  - high-throughput
+authors:
+  - name: Zhiang He
+    affiliation: 1
+    
+affiliations:
+  - name: Independent Researcher
+    index: 1
+date: "2025-04-15"
+bibliography: paper.bib
+repository: https://github.com/HzaCode/ChemInformant
+version: 2.2
+license: MIT
+---
+
+## Summary
+
+ChemInformant is a Python client engineered for programmatic access to PubChem, specifically targeting high-throughput and automated data retrieval tasks. Its architecture streamlines the entire workflow from data acquisition to analysis by directly converting large, mixed-type lists of chemical identifiers into analysis-ready Pandas DataFrames [@Pandas]. To ensure operational resilience, the package natively integrates a suite of robustness features, including persistent HTTP caching, automatic rate-limiting with exponential backoff retries, and runtime data validation using Pydantic [@Pydantic]. By systematically addressing critical limitations in existing tools—such as network instability and inefficient batch processing—and offering up to a 48-fold performance increase, ChemInformant delivers a significantly more reliable and efficient component for the modern Python cheminformatics ecosystem.
+
+## Statement of Need
+
+As these workflows become increasingly automated and scaled, researchers encounter recurring challenges with existing client libraries, primarily concerning network reliability, batch processing capabilities, and a lack of workflow-centric API design.
+
+First, network stability is a significant operational concern. The PubChem API service [@Kim2018PUGREST] enforces dynamic rate limits (e.g., ≤5 requests per second) and may return HTTP 503 (Server Busy) errors during periods of high traffic [@PubChemUsagePolicy]. Many existing clients, such as PubChemPy [@PubChemPy], do not include built-in mechanisms for automatic request throttling or exponential backoff retries. This can lead to script fragility in automated environments, often requiring users to implement manual delays. Furthermore, the general absence of a persistent caching layer results in redundant network requests for repeated queries, which increases latency and unnecessarily consumes API usage quotas.
+
+Second, limitations in handling heterogeneous inputs and providing clear error feedback for batch operations create inefficiencies in high-throughput data processing. Scientific workflows often involve large lists of mixed-type identifiers (e.g., a combination of names, CIDs, and SMILES). Typically, existing tools require users to pre-process these lists into homogeneous groups, adding a preparatory step to the workflow. Additionally, their fault tolerance for batch queries can be limited; a single invalid identifier may cause an entire operation to fail or return incomplete data without explicitly indicating which inputs were problematic. The lack of structured partial success and failure reporting complicates error diagnostics and can affect the reliability of data acquisition pipelines.
+
[…]
