+++ tests/test_cheminfo_api_unit.py
@@ -1,0 +1,227 @@
+from unittest import mock
+import pytest
+import pandas as pd
+
+from src.ChemInformant import cheminfo_api
+from src.ChemInformant.models import Compound, NotFoundError, AmbiguousIdentifierError
+
+
+class TestResolveCid:
+
+    def test_resolve_integer_cid(self):
+        result = cheminfo_api._resolve_to_single_cid(2519)
+        assert result == 2519
+
+    def test_resolve_string_digit_cid(self):
+        result = cheminfo_api._resolve_to_single_cid("2519")
+        assert result == 2519
+
+    def test_resolve_negative_cid_raises_error(self):
+        with pytest.raises(ValueError):
+            cheminfo_api._resolve_to_single_cid(-1)
+
+    def test_resolve_zero_cid_raises_error(self):
+        with pytest.raises(ValueError):
+            cheminfo_api._resolve_to_single_cid(0)
+
+    def test_resolve_name_single_result(self):
+        with mock.patch('src.ChemInformant.api_helpers.get_cids_by_name', return_value=[2519]):
+            result = cheminfo_api._resolve_to_single_cid("caffeine")
+            assert result == 2519
+
+    def test_resolve_name_multiple_results_raises_ambiguous(self):
+        with mock.patch('src.ChemInformant.api_helpers.get_cids_by_name', return_value=[2519, 2520]):
+            with pytest.raises(AmbiguousIdentifierError):
+                cheminfo_api._resolve_to_single_cid("ambiguous_name")
+
+    def test_resolve_name_no_results_try_smiles(self):
+        with mock.patch('src.ChemInformant.api_helpers.get_cids_by_name', return_value=None):
+            with mock.patch.object(cheminfo_api, '_looks_like_smiles', return_value=True):
+                with mock.patch('src.ChemInformant.api_helpers.get_cids_by_smiles', return_value=[2519]):
+                    result = cheminfo_api._resolve_to_single_cid("CC(=O)OC1=CC=CC=C1C(=O)O")
+                    assert result == 2519
+
+    def test_resolve_smiles_multiple_results_raises_ambiguous(self):
+        with mock.patch('src.ChemInformant.api_helpers.get_cids_by_name', return_value=None):
+            with mock.patch.object(cheminfo_api, '_looks_like_smiles', return_value=True):
+                with mock.patch('src.ChemInformant.api_helpers.get_cids_by_smiles', return_value=[2519, 2520]):
+                    with pytest.raises(AmbiguousIdentifierError):
+                        cheminfo_api._resolve_to_single_cid("CC(=O)OC1=CC=CC=C1C(=O)O")
+
+    def test_resolve_not_found_raises_error(self):
+        with mock.patch('src.ChemInformant.api_helpers.get_cids_by_name', return_value=None):
+            with mock.patch.object(cheminfo_api, '_looks_like_smiles', return_value=False):
+                with pytest.raises(NotFoundError):
+                    cheminfo_api._resolve_to_single_cid("unknown_compound")
+
+
+class TestLooksLikeSmiles:
+
+    def test_simple_smiles(self):
+        assert cheminfo_api._looks_like_smiles("C1CCCCC1") == True
+        assert cheminfo_api._looks_like_smiles("C(=O)O") == True
+        assert cheminfo_api._looks_like_smiles("CCBr") == True
+
+    def test_aromatic_smiles(self):
+        assert cheminfo_api._looks_like_smiles("c1ccccc1") == True
+        assert cheminfo_api._looks_like_smiles("c1ccc(O)cc1") == True
+
+    def test_branched_smiles(self):
+        assert cheminfo_api._looks_like_smiles("CC(C)C") == True
[…]
