+++ tests/test_api_helpers_unit.py
@@ -1,0 +1,250 @@
+import time
+import tempfile
+from pathlib import Path
+from unittest import mock
+import pytest
+import requests
+import requests_cache
+
+from src.ChemInformant import api_helpers
+
+
+class TestCacheSetup:
+
+    def test_setup_cache_with_defaults(self):
+        api_helpers.setup_cache()
+        session = api_helpers.get_session()
+        assert isinstance(session, requests_cache.CachedSession)
+        
+    def test_setup_cache_with_custom_params(self):
+        with tempfile.TemporaryDirectory() as tmp_dir:
+            cache_path = Path(tmp_dir) / "test_cache"
+            api_helpers.setup_cache(
+                cache_name=str(cache_path),
+                backend="sqlite",
+                expire_after=3600
+            )
+            session = api_helpers.get_session()
+            assert isinstance(session, requests_cache.CachedSession)
+            if hasattr(session.cache, 'close'):
+                session.cache.close()
+            api_helpers._session = None
+
+    def test_get_session_initializes_when_none(self):
+        api_helpers._session = None
+        session = api_helpers.get_session()
+        assert isinstance(session, requests_cache.CachedSession)
+
+
+class TestNetworkRequests:
+
+    def test_execute_fetch(self):
+        with mock.patch.object(api_helpers, 'get_session') as mock_get_session:
+            mock_session = mock.Mock()
+            mock_response = mock.Mock()
+            mock_session.get.return_value = mock_response
+            mock_get_session.return_value = mock_session
+            
+            result = api_helpers._execute_fetch("http://example.com")
+            
+            assert result == mock_response
+            mock_session.get.assert_called_once_with("http://example.com", timeout=api_helpers.REQUEST_TIMEOUT)
+
+    def test_fetch_with_successful_response(self):
+        mock_response = mock.Mock()
+        mock_response.status_code = 200
+        mock_response.headers = {'Content-Type': 'application/json'}
+        mock_response.json.return_value = {'test': 'data'}
+        
+        with mock.patch.object(api_helpers, '_execute_fetch', return_value=mock_response):
+            result = api_helpers._fetch_with_ratelimit_and_retry("http://example.com")
+            assert result == {'test': 'data'}
+
+    def test_fetch_with_404_response(self):
+        mock_response = mock.Mock()
+        mock_response.status_code = 404
+        
+        with mock.patch.object(api_helpers, '_execute_fetch', return_value=mock_response):
+            result = api_helpers._fetch_with_ratelimit_and_retry("http://example.com")
+            assert result is None
+
+    def test_fetch_with_503_response_and_retry(self):
+        mock_response_503 = mock.Mock()
+        mock_response_503.status_code = 503
+        mock_response_503.from_cache = False
+        
+        mock_response_200 = mock.Mock()
+        mock_response_200.status_code = 200
+        mock_response_200.headers = {'Content-Type': 'application/json'}
+        mock_response_200.json.return_value = {'data': 'success'}
+        
+        with mock.patch.object(api_helpers, '_execute_fetch', side_effect=[mock_response_503, mock_response_200]):
+            with mock.patch('time.sleep'):
+                result = api_helpers._fetch_with_ratelimit_and_retry("http://example.com")
+                assert result == {'data': 'success'}
+
+    def test_fetch_with_cached_503_response(self):
+        mock_response_503 = mock.Mock()
+        mock_response_503.status_code = 503
+        mock_response_503.from_cache = True
+        mock_response_503.request = mock.Mock()
+        
+        mock_response_200 = mock.Mock()
+        mock_response_200.status_code = 200
+        mock_response_200.headers = {'Content-Type': 'application/json'}
[…]

[…]
