+++ tests/test_cli_unit.py
@@ -7,8 +7,8 @@ import argparse
 import pytest
 import pandas as pd
 
-from src.ChemInformant import cli
-from src.ChemInformant.models import NotFoundError
+from ChemInformant import cli
+from ChemInformant.models import NotFoundError
 
 
 class TestMainFetch:
@@ -23,7 +23,7 @@ class TestMainFetch:
         })
         
         with mock.patch('sys.argv', ['chemfetch', 'caffeine', '--props', 'cas,molecular_weight']):
-            with mock.patch('src.ChemInformant.cli.get_properties', return_value=test_data):
+            with mock.patch('ChemInformant.cli.get_properties', return_value=test_data):
                 with mock.patch('builtins.print') as mock_print:
                     cli.main_fetch()
                     mock_print.assert_called()
@@ -37,7 +37,7 @@ class TestMainFetch:
         })
         
         with mock.patch('sys.argv', ['chemfetch', 'caffeine', '--format', 'csv']):
-            with mock.patch('src.ChemInformant.cli.get_properties', return_value=test_data):
+            with mock.patch('ChemInformant.cli.get_properties', return_value=test_data):
                 with mock.patch('builtins.print') as mock_print:
                     cli.main_fetch()
                     mock_print.assert_called()
@@ -52,7 +52,7 @@ class TestMainFetch:
         })
         
         with mock.patch('sys.argv', ['chemfetch', 'caffeine', '--format', 'json']):
-            with mock.patch('src.ChemInformant.cli.get_properties', return_value=test_data):
+            with mock.patch('ChemInformant.cli.get_properties', return_value=test_data):
                 with mock.patch('builtins.print') as mock_print:
                     cli.main_fetch()
                     mock_print.assert_called()
@@ -68,8 +68,8 @@ class TestMainFetch:
             tmp_path = tmp_file.name
         
         with mock.patch('sys.argv', ['chemfetch', 'caffeine', '--format', 'sql', '-o', tmp_path]):
-            with mock.patch('src.ChemInformant.cli.get_properties', return_value=test_data):
-                with mock.patch('src.ChemInformant.cli.df_to_sql') as mock_df_to_sql:
+            with mock.patch('ChemInformant.cli.get_properties', return_value=test_data):
+                with mock.patch('ChemInformant.cli.df_to_sql') as mock_df_to_sql:
                     with mock.patch('builtins.print') as mock_print:
                         cli.main_fetch()
                         mock_df_to_sql.assert_called_once()
@@ -84,14 +84,14 @@ class TestMainFetch:
         empty_df = pd.DataFrame()
         
         with mock.patch('sys.argv', ['chemfetch', 'caffeine']):
-            with mock.patch('src.ChemInformant.cli.get_properties', return_value=empty_df):
+            with mock.patch('ChemInformant.cli.get_properties', return_value=empty_df):
                 with mock.patch('builtins.print') as mock_print:
                     cli.main_fetch()
                     assert mock_print.called
 
     def test_main_fetch_value_error_handling(self):
         with mock.patch('sys.argv', ['chemfetch', 'caffeine']):
-            with mock.patch('src.ChemInformant.cli.get_properties', side_effect=ValueError("Invalid property")):
+            with mock.patch('ChemInformant.cli.get_properties', side_effect=ValueError("Invalid property")):
                 with mock.patch('builtins.print') as mock_print:
                     with pytest.raises(SystemExit) as exc_info:
                         cli.main_fetch()
@@ -100,7 +100,7 @@ class TestMainFetch:
 
     def test_main_fetch_general_exception_handling(self):
         with mock.patch('sys.argv', ['chemfetch', 'caffeine']):
[…]
