### 2025-07-26 0d94ea6
Update coverage badge [skip ci] b993f67e6195a15c4b05b719bd8354de7b231994
