+++ tests/test_cli_unit.py
@@ -1,0 +1,185 @@
+import io
+import sys
+import tempfile
+from pathlib import Path
+from unittest import mock
+import argparse
+import pytest
+import pandas as pd
+
+from ChemInformant import cli
+from ChemInformant.models import NotFoundError
+
+
+class TestMainFetch:
+
+    def test_main_fetch_with_table_format(self):
+        test_data = pd.DataFrame({
+            'input_identifier': ['caffeine'],
+            'cid': ['2519'],
+            'status': ['OK'],
+            'cas': ['58-08-2'],
+            'molecular_weight': ['194.19']
+        })
+        
+        with mock.patch('sys.argv', ['chemfetch', 'caffeine', '--props', 'cas,molecular_weight']):
+            with mock.patch('ChemInformant.cli.get_properties', return_value=test_data):
+                with mock.patch('builtins.print') as mock_print:
+                    cli.main_fetch()
+                    mock_print.assert_called()
+
+    def test_main_fetch_with_csv_format(self):
+        test_data = pd.DataFrame({
+            'input_identifier': ['caffeine'],
+            'cid': ['2519'],
+            'status': ['OK'],
+            'cas': ['58-08-2']
+        })
+        
+        with mock.patch('sys.argv', ['chemfetch', 'caffeine', '--format', 'csv']):
+            with mock.patch('ChemInformant.cli.get_properties', return_value=test_data):
+                with mock.patch('builtins.print') as mock_print:
+                    cli.main_fetch()
+                    mock_print.assert_called()
+                    call_args = str(mock_print.call_args)
+                    assert 'caffeine' in call_args
+
+    def test_main_fetch_with_json_format(self):
+        test_data = pd.DataFrame({
+            'input_identifier': ['caffeine'],
+            'cid': ['2519'],
+            'status': ['OK']
+        })
+        
+        with mock.patch('sys.argv', ['chemfetch', 'caffeine', '--format', 'json']):
+            with mock.patch('ChemInformant.cli.get_properties', return_value=test_data):
+                with mock.patch('builtins.print') as mock_print:
+                    cli.main_fetch()
+                    mock_print.assert_called()
+
+    def test_main_fetch_with_sql_format(self):
+        test_data = pd.DataFrame({
+            'input_identifier': ['caffeine'],
+            'cid': ['2519'],
+            'status': ['OK']
+        })
+        
+        with tempfile.NamedTemporaryFile(suffix='.db', delete=False) as tmp_file:
+            tmp_path = tmp_file.name
+        
+        with mock.patch('sys.argv', ['chemfetch', 'caffeine', '--format', 'sql', '-o', tmp_path]):
+            with mock.patch('ChemInformant.cli.get_properties', return_value=test_data):
+                with mock.patch('ChemInformant.cli.df_to_sql') as mock_df_to_sql:
+                    with mock.patch('builtins.print') as mock_print:
+                        cli.main_fetch()
+                        mock_df_to_sql.assert_called_once()
+                        mock_print.assert_called()
+
+    def test_main_fetch_sql_without_output_fails(self):
+        with mock.patch('sys.argv', ['chemfetch', 'caffeine', '--format', 'sql']):
+            with pytest.raises(SystemExit):
+                cli.main_fetch()
+
+    def test_main_fetch_empty_dataframe(self):
+        empty_df = pd.DataFrame()
+        
+        with mock.patch('sys.argv', ['chemfetch', 'caffeine']):
+            with mock.patch('ChemInformant.cli.get_properties', return_value=empty_df):
+                with mock.patch('builtins.print') as mock_print:
+                    cli.main_fetch()
+                    assert mock_print.called
+
+    def test_main_fetch_value_error_handling(self):
[…]
