### 2025-07-27 951212e
Update coverage badge [skip ci] bf9495f2b083fc8a32aacfe8c1d7983c61304074
