### 2025-07-27 43f0744
Update coverage badge [skip ci] 807afa353855b2c4b6234fb012033db55b4820b8
