### 2025-07-27 #8341 @HzaCode
Hi @espottesmith,

I've now set the version to `v2.3.0` and the archive to `10.5281/zenodo.16482964`.

Please let me know if you need anything else from me at this stage.

Thanks!
