### 2025-07-27 1ff950b
Update coverage badge [skip ci] 5a1a02d2ae5d3324cff2c8852a1ce5aec30e5d00
