### 2025-07-30 #8341 @HzaCode
Hi @LIVazquezS,

Thank you very much for your constructive feedback and positive assessment! I agree that your suggestions for the `get_properties` function will significantly improve its flexibility.

I will implement the following changes:

1.  Make the `properties` parameter optional, returning a default set of core properties when it's not specified.
2.  Expand the function to allow users to retrieve a much wider range of properties available from PubChem.

I will begin working on this right away and will notify you as soon as the updates are complete.

Thanks again for your valuable guidance!
