+++ src/ChemInformant/api_helpers.py
@@ -156,10 +156,24 @@ def get_cids_by_name(name: str) -> List[int] | None:
     """
     Fetches PubChem Compound IDs (CIDs) for a given chemical name.
 
-    :param name: The chemical name to search for (e.g., "aspirin").
-    :type name: str
-    :return: A list of integer CIDs matching the name, or ``None`` if not found.
-    :rtype: List[int] | None
+    This function searches PubChem's database for compounds matching the provided
+    chemical name. It can return multiple CIDs if the name matches multiple compounds.
+
+    Args:
+        name: The chemical name to search for (e.g., "aspirin", "acetylsalicylic acid")
+
+    Returns:
+        List of integer CIDs matching the name, or None if not found
+
+    Examples:
+        >>> get_cids_by_name("aspirin")
+        [2244]
+        >>> get_cids_by_name("glucose") 
+        [5793, 64689, ...]  # Multiple isomers/forms
+
+    Note:
+        This function is used internally by get_properties() for name-to-CID resolution.
+        End users should typically use get_properties() instead.
     """
     url  = f"{PUBCHEM_API_BASE}/compound/name/{quote(name)}/cids/JSON"
     data = _fetch_with_ratelimit_and_retry(url)
@@ -169,10 +183,26 @@ def get_cids_by_smiles(smiles: str) -> List[int] | None:
     """
     Fetches PubChem Compound IDs (CIDs) for a given SMILES string.
 
-    :param smiles: The SMILES string representing the molecule.
-    :type smiles: str
-    :return: A list of integer CIDs matching the SMILES, or ``None`` if not found.
-    :rtype: List[int] | None
+    This function searches PubChem for compounds with structures matching the
+    provided SMILES representation. May return multiple CIDs for stereoisomers
+    or different representations of the same molecule.
+
+    Args:
+        smiles: The SMILES string representing the molecule
+                (e.g., "CC(=O)OC1=CC=CC=C1C(=O)O" for aspirin)
+
+    Returns:
+        List of integer CIDs matching the SMILES, or None if not found
+
+    Examples:
+        >>> get_cids_by_smiles("CC(=O)OC1=CC=CC=C1C(=O)O")
+        [2244]
+        >>> get_cids_by_smiles("CCO")  # Ethanol
+        [702]
+
+    Note:
+        This function is used internally by get_properties() for SMILES-to-CID resolution.
+        End users should typically use get_properties() instead.
     """
     url  = f"{PUBCHEM_API_BASE}/compound/smiles/{quote(smiles)}/cids/JSON"
     data = _fetch_with_ratelimit_and_retry(url)
@@ -183,13 +213,31 @@ def get_batch_properties(cids: List[int], props: List[str]) -> Dict[int, Dict[st
     Fetches multiple properties for a batch of CIDs in a single request,
     handling API pagination automatically.
 
-    :param cids: A list of CIDs to query.
-    :type cids: List[int]
-    :param props: A list of property names to retrieve (e.g., "MolecularWeight", "XLogP").
-    :type props: List[str]
-    :return: A dictionary mapping each requested CID to its properties. If a CID
-             was not found or had no properties, it will map to an empty dict.
-    :rtype: Dict[int, Dict[str, Any]]
+    This is the core function for efficient bulk property retrieval from PubChem.
+    It automatically handles API pagination when dealing with large batches and
+    includes rate limiting and retry logic for reliable data fetching.
+
+    Args:
+        cids: List of compound IDs to query
+        props: List of property names in CamelCase format 
+               (e.g., ["MolecularWeight", "XLogP", "CanonicalSMILES"])
+               Must use exact PubChem API property names
+
+    Returns:
[…]
