+++ src/ChemInformant/cheminfo_api.py
@@ -1,247 +1,828 @@
 """
 This module contains the high-level, user-facing API for ChemInformant.
-These functions are designed for ease of use and to support common
-cheminformatics workflows.
+It now operates exclusively on a snake_case standard for all property names.
 """
-
 from __future__ import annotations
-
 import re
 import sys
-from typing import Iterable, List, Union, Dict, Any
-
+import logging
+from typing import Iterable, List, Union, Dict, Any, Optional
 import pandas as pd
-
 from . import api_helpers
 from .models import Compound, NotFoundError, AmbiguousIdentifierError
+from .constants import (
+    CORE_PROPS, THREED_PROPS, ALL_PROPS, PROPERTY_ALIASES, SNAKE_TO_CAMEL, CAMEL_TO_SNAKE
+)
 
 _SMILES_TOKENS = re.compile(r"[=#\[\]\(\)]|\d|Br|Cl|Si", re.I)
-
+_SPECIAL_PROPS = {"cas", "synonyms"}
 
 def _looks_like_smiles(s: str) -> bool:
-    """Checks if a string contains characters typical of SMILES."""
     return bool(_SMILES_TOKENS.search(s))
 
-
 def _resolve_to_single_cid(identifier: Union[str, int]) -> int:
-    """
-    Resolves a given identifier (name, CID, or SMILES) to a single,
-    unambiguous PubChem CID.
-    """
+    # This function remains robust as is.
     if isinstance(identifier, int):
-        if identifier > 0:
-            return identifier
+        if identifier > 0: return identifier
         raise ValueError(f"Invalid CID: {identifier}")
     if isinstance(identifier, str) and identifier.isdigit():
         return int(identifier)
-
     if isinstance(identifier, str):
         cids = api_helpers.get_cids_by_name(identifier)
         if cids:
-            if len(cids) == 1:
-                return cids[0]
+            if len(cids) == 1: return cids[0]
             raise AmbiguousIdentifierError(identifier, cids)
-
     if isinstance(identifier, str) and _looks_like_smiles(identifier):
         cids = api_helpers.get_cids_by_smiles(identifier)
         if cids:
-            if len(cids) == 1:
-                return cids[0]
+            if len(cids) == 1: return cids[0]
             raise AmbiguousIdentifierError(identifier, cids)
-
     raise NotFoundError(identifier)
 
-
-PROPERTY_ALIASES: Dict[str, List[str]] = {
-    "molecular_weight": ["MolecularWeight"],
-    "molecular_formula": ["MolecularFormula"],
-    "canonical_smiles": ["CanonicalSMILES", "ConnectivitySMILES"],
-    "isomeric_smiles": ["IsomericSMILES", "SMILES"],
-    "iupac_name": ["IUPACName"],
-    "xlogp": ["XLogP"],
-    "inchi": ["InChI"]
-}
-_SPECIAL_PROPS = {"cas", "synonyms"}
-
-
 def get_properties(
-    identifiers: Iterable[Union[str, int]],
-    properties: Iterable[str],
+    identifiers: Union[int, str, List[int], List[str]],
+    properties: Optional[Union[str, List[str]]] = None,
+    *,
+    namespace: str = "cid",
+    include_3d: bool = False,
+    all_properties: bool = False,
+    **kwargs,
 ) -> pd.DataFrame:
     """
-    Retrieves multiple properties for a list of chemical identifiers.
+    Retrieve chemical properties for one or more compounds from PubChem.
+    
+    This function is the core interface for fetching molecular properties. It accepts
+    various types of chemical identifiers and returns data in a standardized snake_case
+    format with consistent column ordering and error handling.
+    
+    Args:
[…]
