### 2025-07-31 #8341 @HzaCode
Hi @espottesmith, @LIVazquezS,

I have completed the updates based on **@LIVazquezS's** feedback. Following your suggestions, I've made changes to both the convenience functions and the `get_properties` function to improve their utility and flexibility.

Here is a summary of the changes:

*   **Expanded Convenience Functions:** The number of standalone convenience functions has been increased from 8 to 22. This provides more direct, single-purpose functions for retrieving common properties like InChI/InChIKey, stereochemistry counts, and mass spectrometry data.

*   **Revised `get_properties` Function:** As you recommended, the `get_properties` function has been updated to be more flexible. It now supports four distinct calling modes, each corresponding to a specific use case:
    *   **Mode 1 (Default):** The `properties` parameter is now optional. Calling `get_properties(identifiers)` returns a default set of 22 "CORE" properties.
    *   **Mode 2 (Custom):** Users can retrieve any combination of the 40+ available attributes by passing a list or string to the `properties` parameter, for example: `get_properties(identifiers, properties=['molecular_weight', 'xlogp'])`.
    *   **Mode 3 (3D Data):** The boolean flag `include_3d=True` can be used to retrieve the CORE properties plus the 14 available 3D properties. The call is `get_properties(identifiers, include_3d=True)`.
    *   **Mode 4 (All Data):** The `all_properties=True` flag provides a simple way to fetch all available properties from PubChem via the call `get_properties(identifiers, all_properties=True)`.

The related documentation and tests have also been updated to reflect these changes.

The diagram below outlines the new structure, showing how the API now addresses both simple queries (via the expanded convenience functions) and advanced queries (via the revised `get_properties` function).

Thank you again for your constructive guidance, which has helped improve the package.
