### 2025-07-31 0bbc988
Update coverage badge [skip ci] 927e4441fedb7c0bac06c56a5eeee5c62b77e041
