### 2025-07-31 #8341 @HzaCode
To reflect these changes in the software package, I have released version `v2.4.0` and archived it on Zenodo with the corresponding `DOI`: [`10.5281/zenodo.16623785`](https://doi.org/10.5281/zenodo.16623785).

Please let me know if any further information is required from my side.

Thanks!
