### 2025-07-31 #8341 @LIVazquezS
Hi @HzaCode,

Thanks for your detailed and prompt answer. All my points have been addressed. Your code would become an essential tool for the cheminformatics community.

I am looking forward to using it in my research.

@espottesmith, please consider my review complete.  I am happy to recommend this work for acceptance in JOSS.

Regards,

L
