### 2025-08-01 #8341 @HzaCode
Thanks for running the final checks.

I've addressed the license issue by removing the non-compliant MIT license file.
