### 2025-08-01 d06ede8
Update paper.md
