+++ paper/paper.md
@@ -25,25 +25,22 @@ version: 2.4.0
 license: MIT
 ---
 
+
 ## Summary
 
-ChemInformant is a Python client engineered for programmatic access to PubChem, specifically targeting high-throughput and automated data retrieval tasks. Its architecture streamlines the entire workflow from data acquisition to analysis by directly converting large, mixed-type lists of chemical identifiers into analysis-ready Pandas DataFrames [@Pandas]. To ensure operational resilience, the package natively integrates a suite of robustness features, including persistent HTTP caching, automatic rate-limiting with exponential backoff retries, and runtime data validation using Pydantic [@Pydantic]. By systematically addressing critical limitations in existing tools—such as network instability and inefficient batch processing—and offering up to a 48-fold performance increase, ChemInformant delivers a significantly more reliable and efficient component for the modern Python cheminformatics ecosystem.
+ChemInformant is a Python client for high-throughput, programmatic access to PubChem. It streamlines automated data retrieval by converting large, mixed-type lists of chemical identifiers directly into analysis-ready Pandas DataFrames [@Pandas]. To ensure resilience, the package integrates persistent HTTP caching, automatic rate-limiting with exponential backoff retries, and runtime data validation using Pydantic [@Pydantic]. By addressing critical limitations in existing tools, such as network instability and inefficient batch processing, ChemInformant offers up to a 48-fold performance increase, providing a more reliable and efficient component for the modern Python cheminformatics ecosystem.
 
 ## Statement of Need
 
-As these workflows become increasingly automated and scaled, researchers encounter recurring challenges with existing client libraries, primarily concerning network reliability, batch processing capabilities, and a lack of workflow-centric API design.
+Automated cheminformatics workflows require robust and efficient data access, but researchers face recurring challenges with existing PubChem clients. Network reliability is a primary concern. The PubChem API enforces dynamic rate limits, which can halt automated scripts [@PubChemUsagePolicy]. Many clients, like PubChemPy [@PubChemPy], lack built-in request throttling, retries, or persistent caching, forcing users to implement boilerplate code to handle network errors and redundant requests.
 
-First, network stability is a significant operational concern. The PubChem API service [@Kim2018PUGREST] enforces dynamic rate limits (e.g., ≤5 requests per second) and may return HTTP 503 (Server Busy) errors during periods of high traffic [@PubChemUsagePolicy]. Many existing clients, such as PubChemPy [@PubChemPy], do not include built-in mechanisms for automatic request throttling or exponential backoff retries. This can lead to script fragility in automated environments, often requiring users to implement manual delays. Furthermore, the general absence of a persistent caching layer results in redundant network requests for repeated queries, which increases latency and unnecessarily consumes API usage quotas.
+Batch processing is also often inefficient. Workflows with mixed-type identifiers (e.g., names and CIDs) require manual pre-processing. Furthermore, a single invalid identifier in a large batch can cause an entire query to fail without clear error reporting, hindering data acquisition pipelines.
 
[…]
