### 2025-08-01 b09d571
Update coverage badge [skip ci] d1ec1a497e668ed86e70b7487e2aa1430a035de3
