### 2025-08-01 6707522
Delete LICENSE.md
