### 2025-08-01 7da28ed
Update coverage badge [skip ci] 8efb3bb575a3962cb1fc2a489d0034bcb57ac89b
