### 2025-08-01 #8341 @HzaCode
Thank you for your guidance. I have completed all the required post-review tasks.

- [ ] All authors, affiliations, and ORCIDs in the paper have been double-checked.
- [ ] A new release of the software, version **v2.4.0**, has been published to reflect the latest changes.
- [ ] The new release has been archived on Zenodo. The DOI is: **[10.5281/zenodo.16623785](https://doi.org/10.5281/zenodo.16623785)**.
- [ ]  I have confirmed that the title and author list in the Zenodo archive match those in the JOSS paper.
- [ ] I have confirmed that the license listed in the Zenodo archive is the same as the software's license.

Please let me know if anything else is needed from my side.
