### 2025-08-04 #8341 @espottesmith
References look good.

Paper comments:
- "By addressing critical limitations in existing tools, such as network instability and inefficient
batch processing, ChemInformant offers up to a 48-fold performance increase, providing a
more reliable and efficient component for the modern Python cheminformatics ecosystem" <-- 48-fold performance increase relative to what?

Otherwise, this looks quite good!
