### 2025-08-04 e2247b5
Update coverage badge [skip ci] 15928094beb73ef2754c0f713012ef8644f1fe6a
