### 2025-08-04 ed6d0ba
Update paper.md
