### 2025-08-04 fabb579
Update paper.md
