### 2025-08-07 8cb3d67
Update coverage badge [skip ci] c39ca3c1c713e9c77869a0359237264e1da4c011
