### 2025-08-12 29db9a8
Update coverage badge [skip ci] beab989d25e00d20ce3b300b2f377b2c0f1e86e5
